### 2019-01-27 #1205 @whedon
[PRE REVIEW]: DmpBbo: A versatile Python/C++ library for Function Approximation, Dynamical Movement Primitives, and Black-Box Optimization
**Submitting author:** @graiola (<a href="http://orcid.org/0000-0003-1481-1106">Gennaro Raiola</a>)
**Repository:** <a href="https://github.com/stulp/dmpbbo" target ="_blank">https://github.com/stulp/dmpbbo</a>
**Version:**  v0.7-beta
