### 2019-01-27 #1205 @danielskatz
👋 @giraola - Please help by looking at [this list of potential reviewers](
https://bit.ly/joss-reviewers) and identify individuals who might be able to review your submission (please start at the bottom of the list). Also, please suggest individuals who might be good reviewers but are not on this list, ideally by mentioning their GitHub handles here, but if not, just by name.
