### 2019-01-27 #1205 @danielskatz
Also, note a couple of issues in the paper - a couple of references appear as ????s, indicating that the text in {}s in the .md paper didn't match an entry in the .bib file, and the second reference has some case problems in the proceedings (booktitle) name. In addition, that reference should likely have a DOI.
After you make changes in the .md or .bib files, type ```@whedon generate pdf``` as the first line in a new comment here to rebuild the PDF, then check it.
