### 2019-01-29 0a518d3
Add missing doi and missing references
