+++ docs/dmp.bib
@@ -1,7 +1,6 @@
 % This file was created with JabRef 2.10.
 % Encoding: ISO8859_1
 
-
 @Article{buchli11learning,
   title                    = {Learning Variable Impedance Control},
   author                   = {Jonas Buchli and Freek Stulp and Evangelos Theodorou and Stefan Schaal},
@@ -10,6 +9,7 @@
   number                   = {7},
   pages                    = {820-833},
   volume                   = {30},
+  doi                      = {https://doi.org/10.1177/0278364911402527},
 
   abstract                 = {One of the hallmarks of the performance, versatility, and robustness of biological motor control is the ability to adapt the impedance of the overall biomechanical system to different task requirements and stochastic disturbances. A transfer of this principle to robotics is desirable, for instance to enable robots to work robustly and safely in everyday human environments. It is, however, not trivial to derive variable impedance controllers for practical high degree-of-freedom (DOF) robotic tasks. In this contribution, we accomplish such variable impedance control with the reinforcement learning (RL) algorithm PI2 ({\bf P}olicy {\bf I}mprovement with {\bf P}ath {\bf I}ntegrals). PI2 is a model-free, sampling based learning method derived from first principles of stochastic optimal control. The PI2 algorithm requires no tuning of algorithmic parameters besides the exploration noise. The designer can thus fully focus on cost function design to specify the task. From the viewpoint of robotics, a particular useful property of PI2 is that it can scale to problems of many DOFs, so that reinforcement learning on real robotic systems becomes feasible. We sketch the PI2 algorithm and its theoretical properties, and how it is applied to gain scheduling for variable impedance control. We evaluate our approach by presenting results on several simulated and real robots. We consider tasks involving accurate tracking through via-points, and manipulation tasks requiring physical contact with the environment. In these tasks, the optimal strategy requires both tuning of a reference trajectory \emph{and} the impedance of the end-effector. The results show that we can use path integral based reinforcement learning not only for planning but also to derive variable gain feedback controllers in realistic scenarios. Thus, the power of variable impedance control is made available to a wide variety of robotic systems and practical applications.},
   bib2html_pubtype         = {Journal},
@@ -25,6 +25,7 @@
   number                   = {2},
   pages                    = {328--373},
   volume                   = {25},
+  doi                      = {https://doi.org/10.1162/neco_a_00393},
 
   file                     = {ijspeert13dynamical.pdf:unsorted/ijspeert13dynamical.pdf:PDF},
   key                      = {movement primitives, dynamic systems, rhythmic, discrete, perception-action coupling},
@@ -58,6 +59,7 @@
   number                   = {1},
   pages                    = {145-157},
   volume                   = {28},
+  doi                      = {https://doi.org/10.1109/tro.2011.2163863},
 
   bibsource                = {DBLP, http://dblp.uni-trier.de},
   ee                       = {http://dx.doi.org/10.1109/TRO.2011.2163863},
@@ -72,6 +74,7 @@
   number                   = {5},
   pages                    = {493-500},
   volume                   = {24},
+  doi                      = {https://doi.org/10.1016/j.neunet.2011.02.004},
 
   file                     = {matsubara11learning.pdf:unsorted/matsubara11learning.pdf:PDF}
 }
@@ -114,7 +117,7 @@
   author                   = {Freek Stulp and Laura Herlant and Antoine Hoarau and Gennaro Raiola},
   booktitle                = {International Conference on Intelligent Robots and Systems (IROS)},
   year                     = {2014},
-
+  doi                      = {10.1109/IROS.2014.6942741}, 
   bib2html_pubtype         = {Refereed Conference Paper},
   bib2html_rescat          = {Reinforcement Learning of Robot Skills}
 }
@@ -124,7 +127,7 @@
[…]
