### 2019-01-29 #1205 @danielskatz
👋 @studywolf - would you be willing to review this submission to JOSS?
