### 2019-01-29 #1205 @studywolf
Hello! Looks very interesting, I'd be happy to review. One possible issue is that I am just about to take off for a couple weeks of travelling and probably won't have time before I go. What's the timeline here?
