### 2019-02-01 #1225 @whedon
[REVIEW]: DmpBbo: A versatile Python/C++ library for Function Approximation, Dynamical Movement Primitives, and Black-Box Optimization
**Submitting author:** @graiola (<a href="http://orcid.org/0000-0003-1481-1106">Gennaro Raiola</a>)
**Repository:** <a href="https://github.com/stulp/dmpbbo" target ="_blank">https://github.com/stulp/dmpbbo</a>
**Version:** 1.0.0
