### 2019-02-01 #1225 @danielskatz
👋 @studywolf & @jenskober - thanks for agreeing to review.

I'll note here that we expect a couple of weeks delay until this gets started.

When you are ready to start, please see the comment above - your job is basically to review the paper and software according to your checklists, checking off items as they are deemed ok, and making either short comments or more detailed comments in the code repository as you find problems that need to be addressed by the author.
