### 2019-02-03 9c5b0bb
Removed links to local PDFs
