+++ docs/dmp.bib
@@ -25,8 +25,6 @@
   number                   = {2},
   pages                    = {328--373},
   volume                   = {25},
-
-  file                     = {ijspeert13dynamical.pdf:unsorted/ijspeert13dynamical.pdf:PDF},
   key                      = {movement primitives, dynamic systems, rhythmic, discrete, perception-action coupling},
   url                      = {http://www-clmc.usc.edu/publications/I/ijspeert-NC2013.pdf}
 }
@@ -36,8 +34,6 @@
   author                   = {A. J. Ijspeert and J. Nakanishi and S. Schaal},
   booktitle                = {Proceedings of the IEEE International Conference on Robotics and Automation (ICRA)},
   year                     = {2002},
-
-  file                     = {:/home/stulp/docs/bibliography/robotics/clmc-lab/Ijspeert, Nakanishi, Schaal - 2002 - Movement Imitation with Nonlinear Dynamical Systems in Humanoid Robots.pdf:PDF},
   keywords                 = {clmc lab}
 }
 
@@ -61,7 +57,6 @@
 
   bibsource                = {DBLP, http://dblp.uni-trier.de},
   ee                       = {http://dx.doi.org/10.1109/TRO.2011.2163863},
-  file                     = {kulvicius12joining.pdf:kulvicius12joining.pdf:PDF}
 }
 
 @Article{matsubara11learning,
@@ -72,8 +67,6 @@
   number                   = {5},
   pages                    = {493-500},
   volume                   = {24},
-
-  file                     = {matsubara11learning.pdf:unsorted/matsubara11learning.pdf:PDF}
 }
 
 @InProceedings{silva12learning,
@@ -88,12 +81,9 @@
   pages                    = {1679--1686},
   publisher                = {Omnipress},
   series                   = {ICML '12},
-
-  file                     = {silva12learning.pdf:unsorted/silva12learning.pdf:PDF},
   isbn                     = {978-1-4503-1285-1},
   keywords                 = {manifolds, meta-parameters},
   location                 = {Edinburgh, Scotland, GB},
-  review                   = {1) train K DMPs on K trajectories 2) learn D lower-dimensional surfaces (ISOMAP) 3) train a classifier that maps from T to which of the D surfaces 4) compute a regression from T to each policy parameter individually for each of the D surfaces (SVM)}
 }
 
 @InProceedings{stulp12adaptive,
