### 2019-03-10 #1225 @studywolf
regarding the authorship checkbox, Gennaro Raiola has contributed about 1k vs 120k lines of code from Freek Stulp according to the commit history. I'm not sure exactly what the requirements for 'major contribution' are, but the authorship on the paper reflects Freek Stulp as primary author so I'm checking this box as complete.
