### 2019-03-10 stulp/dmpbbo#28 @studywolf
what versions of python are supported?
Did not see in the install doc: Which versions of Python are supported?
