### 2019-03-10 #1225 @studywolf
didn't see what versions are python are supported. leaving Installation unchecked until stulp/dmpbbo#28 resolved
