### 2019-03-12 5d1b211
Required Python version is python3. Fixes #28.
