+++ docs/dmp.bib
@@ -12,8 +12,6 @@
   doi                      = {https://doi.org/10.1177/0278364911402527},
 
   abstract                 = {One of the hallmarks of the performance, versatility, and robustness of biological motor control is the ability to adapt the impedance of the overall biomechanical system to different task requirements and stochastic disturbances. A transfer of this principle to robotics is desirable, for instance to enable robots to work robustly and safely in everyday human environments. It is, however, not trivial to derive variable impedance controllers for practical high degree-of-freedom (DOF) robotic tasks. In this contribution, we accomplish such variable impedance control with the reinforcement learning (RL) algorithm PI2 ({\bf P}olicy {\bf I}mprovement with {\bf P}ath {\bf I}ntegrals). PI2 is a model-free, sampling based learning method derived from first principles of stochastic optimal control. The PI2 algorithm requires no tuning of algorithmic parameters besides the exploration noise. The designer can thus fully focus on cost function design to specify the task. From the viewpoint of robotics, a particular useful property of PI2 is that it can scale to problems of many DOFs, so that reinforcement learning on real robotic systems becomes feasible. We sketch the PI2 algorithm and its theoretical properties, and how it is applied to gain scheduling for variable impedance control. We evaluate our approach by presenting results on several simulated and real robots. We consider tasks involving accurate tracking through via-points, and manipulation tasks requiring physical contact with the environment. In these tasks, the optimal strategy requires both tuning of a reference trajectory \emph{and} the impedance of the end-effector. The results show that we can use path integral based reinforcement learning not only for planning but also to derive variable gain feedback controllers in realistic scenarios. Thus, the power of variable impedance control is made available to a wide variety of robotic systems and practical applications.},
-  bib2html_pubtype         = {Journal},
-  bib2html_rescat          = {Reinforcement Learning of Variable Impedance Control},
   url                      = {http://ijr.sagepub.com/content/early/2011/03/31/0278364911402527}
 }
 
@@ -25,14 +23,7 @@
   number                   = {2},
   pages                    = {328--373},
   volume                   = {25},
-<<<<<<< HEAD
-=======
   doi                      = {https://doi.org/10.1162/neco_a_00393},
-
-  file                     = {ijspeert13dynamical.pdf:unsorted/ijspeert13dynamical.pdf:PDF},
->>>>>>> 0a518d3df94eafe4ae8c1962279bfb8cd0e499c8
-  key                      = {movement primitives, dynamic systems, rhythmic, discrete, perception-action coupling},
-  url                      = {http://www-clmc.usc.edu/publications/I/ijspeert-NC2013.pdf}
 }
 
 @InProceedings{ijspeert02movement,
@@ -40,7 +31,6 @@
   author                   = {A. J. Ijspeert and J. Nakanishi and S. Schaal},
   booktitle                = {Proceedings of the IEEE International Conference on Robotics and Automation (ICRA)},
   year                     = {2002},
-  keywords                 = {clmc lab}
 }
 
 @InProceedings{kalakrishnan11learning,
@@ -74,12 +64,7 @@
   number                   = {5},
   pages                    = {493-500},
   volume                   = {24},
-<<<<<<< HEAD
-=======
   doi                      = {https://doi.org/10.1016/j.neunet.2011.02.004},
-
-  file                     = {matsubara11learning.pdf:unsorted/matsubara11learning.pdf:PDF}
->>>>>>> 0a518d3df94eafe4ae8c1962279bfb8cd0e499c8
 }
 
 @InProceedings{silva12learning,
@@ -95,7 +80,6 @@
   publisher                = {Omnipress},
   series                   = {ICML '12},
   isbn                     = {978-1-4503-1285-1},
-  keywords                 = {manifolds, meta-parameters},
   location                 = {Edinburgh, Scotland, GB},
 }
 
[…]
