### 2019-03-12 0bf1033
Ubuntu default install of Eigen ist now >= 3.2. Updated doxygen link.
