### 2019-03-12 d60b68d
numpy and matplotlib requirements
