### 2019-03-13 #1225 @graiola
@studywolf, the issue stulp/dmpbbo#28 should be fixed now.
