### 2019-03-14 stulp/dmpbbo#29 @jenskober
Installation Requirements
I installed/compiled dmpbbo on a VirtualBox of fresh Ubuntu 16.04 (desktop) install.

Eigen3 is still required but went missing after a recent commit (Ubuntu now contains a high enough version, so custom install is not required) but should be added to apt-get list, i.e., libeigen3-dev

For the Python part tkinter (python3-tk) was not included by default when installing python3 via apt-get but also required.

In general it would be nice to mention which versions of Debian/Ubuntu this has been tested with.

openjournals/joss-reviews#1225
