### 2019-03-14 #1225 @jenskober
@graiola
missing DOIs in paper
https://doi.org/10.1109/ROBOT.2002.1014739
https://doi.org/10.1109/IROS.2012.6385818
