### 2019-03-14 f925055
Update the travis config to use xenial with libeigen3-dev provided by the official source
