+++ INSTALL.txt
@@ -1,14 +1,9 @@
 ________________________________________________________________________________    
-INSTALLING REQUIRED/OPTIONAL PACKAGES FOR DEBIAN (UBUNTU) USERS
+INSTALLING REQUIRED/OPTIONAL PACKAGES FOR DEBIAN (UBUNTU 16.04) USERS
 (for manual installation see "MANUAL INSTALLATION OF REQUIRED/OPTIONAL PACKAGES" below) 
 
-sudo apt-get install cmake libboost-filesystem-dev libboost-system-dev libboost-serialization-dev 
-sudo apt-get install doxygen graphviz 
-
-"sudo apt-get eigen3" currently (February 2014) installs version 3.1.2. However, DmpBbo
-requires version >= 3.2. To get the latest Eigen version:
-  use this PPA: https://launchpad.net/~kalakris/+archive/eigen
-  or do a manual install: see "MANUAL INSTALLATION OF REQUIRED/OPTIONAL PACKAGES" below
+sudo apt-get install cmake libboost-filesystem-dev libboost-system-dev libboost-serialization-dev libeigen3-dev python3-tk
+sudo apt-get install doxygen graphviz
   
 ________________________________________________________________________________    
 BUILDING DMPBBO 
@@ -61,20 +56,7 @@ MANUAL INSTALLATION OF REQUIRED/OPTIONAL PACKAGES
 
 ___________________________________
 CMAKE (required)
-  http://www.cmake.org/cmake/resources/software.html	
-
-___________________________________
-EIGEN (required, version >= 3.2)
-  http://eigen.tuxfamily.org/dox/GettingStarted.html
-  Getting the latest version (thanks to Ryan Lober):
-		sudo apt-get install mercurial
-		"go to whereever you want to start the installation"
-		hg clone https://bitbucket.org/eigen/eigen/
-		cd eigen
-		mkdir build_dir
-		cd build_dir
-		cmake ..
-  	sudo make install
+  http://www.cmake.org/cmake/resources/software.html
   	
 ___________________________________
 BOOST (required modules: filesystem, system, serialization. version >=1.34) 
