### 2019-03-14 stulp/dmpbbo#29 @graiola
I created a PR #30
