### 2019-03-14 stulp/dmpbbo#31 @jenskober
Issues with Demos and Tests
It would be nice to add some instructions on how to call the demos. This is mentioned for demoDynamicalSystems in the tutorial, however the corresponding plotDynamicalSystem.py does not seem to exist.
The Python Wrappers are nice, but it is also not 100% clear how to call them. You need to "make install" and then call them from the original source directory, which I didn't find very intuitive.

The help for semoDynamicalSystems is quite nice, for the dmp demos I'd suggest changing it to the same format (with a complete example of a call)

If you call demoDmpChangeGoal or similar with a relative path for the output, the python files don't work if called from a different location then from which they were generated. E.g.
./demoDmpChangeGoal output/ and then going to the directory where the Python script resides and trying to call is there

Some of the demos cause a core dump if they don't find the correct file rather than existing nicely
I found demoDmpTrainFromTrajFile and demoOptimizationDmp

related note demoFunctionApproxTrainingWrapper.py crashes if lwpr is not installed while the C++ handles that correctly

I also got core dumps for
functionapproximator/tests/testTrainingCompareCppPython
and
dmp_bbo/tests/testTaskViapoint
