### 2019-03-14 7531823
Add missing doi for stulp12adaptive and ijspeert02movement
