+++ docs/dmp.bib
@@ -30,6 +30,7 @@
   title                    = {Movement imitation with nonlinear dynamical systems in humanoid robots},
   author                   = {A. J. Ijspeert and J. Nakanishi and S. Schaal},
   booktitle                = {Proceedings of the IEEE International Conference on Robotics and Automation (ICRA)},
+  doi                      = {https://doi.org/10.1109/ROBOT.2002.1014739},
   year                     = {2002},
 }
 
@@ -88,6 +89,7 @@
   author                   = {Freek Stulp},
   booktitle                = {International Conference on Intelligent Robots and Systems (IROS)},
   year                     = {2012},
+  doi                      = {https://doi.org/10.1109/IROS.2012.6385818},
   pages                    = {1631-1636},
 
   abstract                 = {Most experiments on policy search for robotics focus on isolated tasks, where the experiment is split into two distinct phases: 1)~the learning phase, where the robot learns the task through exploration; 2)~the exploitation phase, where exploration is turned off, and the robot demonstrates its performance on the task it has learned. In this paper, we present an algorithm that enables robots to continually and autonomously alternate between these phases. We do so by combining the `Policy Improvement with Path Integrals' direct reinforcement learning algorithm with the covariance matrix adaptation rule from the `Cross-Entropy Method' optimization algorithm. This integration is possible because both algorithms iteratively update parameters with probability-weighted averaging. A practical advantage of the novel algorithm, called PI2-CMA, is that it alleviates the user from having to manually tune the degree of exploration. We evaluate PI2-CMA's ability to continually and autonomously tune exploration on two tasks.},
