### 2019-03-14 stulp/dmpbbo#29 @jenskober
I would also add that to INSTALL.txt

unrelated note: some of the Python files also require latex for the plotting, e.g., demoDmpWrapper.py
I had to install texlive and texlive-latex-extra
