### 2019-03-14 #1225 @jenskober
and a few little issues with the demos and tests
stulp/dmpbbo#31
