### 2019-03-14 #1225 @graiola
@jenskober I fixed the missing DOIs, thanks.
