### 2019-03-14 41d7daa
Use a bash script to install the required and optional packages
