+++ INSTALL.txt
@@ -2,15 +2,24 @@ ________________________________________________________________________________
 C++: INSTALLING REQUIRED/OPTIONAL PACKAGES FOR DEBIAN (UBUNTU 16.04) USERS
 (for manual installation see "MANUAL INSTALLATION OF REQUIRED/OPTIONAL PACKAGES" below) 
 
-sudo apt-get install cmake libboost-filesystem-dev libboost-system-dev libboost-serialization-dev libeigen3-dev python3-tk
+REQUIRED PACKAGES:
+
+sudo apt-get install cmake libboost-filesystem-dev libboost-system-dev libboost-serialization-dev libeigen3-dev 
+
+OPTIONAL PACKAGES:
+
 sudo apt-get install doxygen graphviz
 
-PYTHON
+PYTHON:
 
 dmpbbo requires python3
 (mainly due to the way packages are imported)
 
 dmpbbo requires numpy (http://www.numpy.org/) and matplotlib (https://matplotlib.org/)
+
+sudo apt-get install python3-numpy python3-tk python3-matplotlib
+
+To automatically install the packages (required and optional) please run install_dependencies.sh
 
 ________________________________________________________________________________    
 C++: BUILDING DMPBBO 
@@ -47,7 +56,6 @@ This is a bug in boost::serialization for versions >=1.44, which can be fixed wi
     https://svn.boost.org/trac/boost/attachment/ticket/4842/void_cast.cpp.patch
 Note that the crash happens towards the very end of the program, just before returning from main() when desctructors are called. So it does not affect functionality, but it is just very irritating that boost crashes and sometimes dumps a core (I use boost version 1.48, maybe the issue is fixed in later versions). If I had known things like this would happen, I would have probably used cereal (http://uscilab.github.io/cereal/) rather than boost::serialization... 
 
-
 ________________________________________________________________________________    
 C++: REAL-TIME DEBUGGING BUILD
 
@@ -57,7 +65,7 @@ To compile a version that has real-time checks built in:
   mkdir -p build_dir_realtime; cd build_dir_realtime; cmake .. -DREALTIME_CHECKS=1 -DCMAKE_BUILD_TYPE=Debug; make VERBOSE=1
 
 Note that it doesn't make sense to set -DREALTIME_CHECKS=1 without -DCMAKE_BUILD_TYPE=Debug, as asserts are only done in debug mode. 
-  
+
 ________________________________________________________________________________    
 C++: MANUAL INSTALLATION OF REQUIRED/OPTIONAL PACKAGES
 
