+++ .travis.yml
@@ -22,8 +22,6 @@ before_install:
 
   # clang 3.4
   - if [ "$CXX" == "clang++" ]; then sudo add-apt-repository -y ppa:h-rayflood/llvm; fi
-
-  - sudo apt-get update -qq
   
 
 install:
@@ -50,8 +48,7 @@ install:
   - if [ "$CXX" == "clang++" ]; then export CXX="clang++-3.4" CC="clang-3.4"; fi
   
 ## Dependencies to download here
-## Install Eigen 3.2.x
-  - sudo apt-get install -qq libboost-all-dev libeigen3-dev
+  - bash ./install_dependencies.sh
   
 script:
   # compile and install
