### 2019-03-14 stulp/dmpbbo#31 @stulp
Hello Jens,
Thank you for your detailed analysis! They will certainly help improve the quality of the code.
I've taken the liberty of splitting your issue into smaller ones (#32 to #40), so that we can assign & conquer. As the other ones are now open, I'm closing this issue.
Cheers,
Freek
