### 2019-03-14 a6f45cb
Included empty trajectory check for Trajectory::duration()
