### 2019-03-14 88e48ef
Avoid core dump in demoDmpTrainFromTrajectoryFile. Provide some --help. Fixes #36.
