### 2019-03-14 c7e8fb6
Clean exit when directory does not exist. Fixes #37
