+++ demos/dmp_bbo/demoOptimizationDmp.cpp
@@ -50,6 +50,7 @@ using namespace DmpBbo;
  * \param[in] n_args Number of arguments
  * \param[in] args Arguments themselves
  * \return Success of exection. 0 if successful.
+ * \todo To focus on the demo code, it would be nice to have the processing of arguments in a separate function.
  */
 int main(int n_args, char* args[])
 {
@@ -66,6 +67,16 @@ int main(int n_args, char* args[])
     directory = string(args[2]);
   }
 
+  if (!boost::filesystem::exists(directory))
+  {
+    cerr << "Directory '" << directory << "' does not exist." << endl;
+    cerr << "HINT: The preferred way to run this demo is by calling ";
+    cerr << "python demos/dmp_bbo/" << (args[0]+2) <<"Wrapper.py, "; // +2 removes leading "./" 
+    cerr << "rather than this binary." << endl; 
+    cerr << "Abort." << endl;
+    return -1;
+  }
+  
   // Make the task
   //VectorXd viapoint = VectorXd::Constant(n_dims,2.0);
   //viapoint[0] = 2.5;
