### 2019-03-14 e1e2f36
Do not plot if not data was generated (e.g. because LWPR was not installed). Fixes #38.
