### 2019-03-14 #1225 @stulp
> and a few little issues with the demos and tests
> [stulp/dmpbbo#31](https://github.com/stulp/dmpbbo/issues/31)

I've taken the liberty of splitting your issue [stulp/dmpbbo#31](https://github.com/stulp/dmpbbo/issues/31) into smaller ones 32-40, so that we can assign & conquer. As the other ones are now open, I've closed [stulp/dmpbbo#31](https://github.com/stulp/dmpbbo/issues/31).
