### 2019-03-15 1be4130
Re-write INSTALL as md file
