+++ install_dependencies.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
 sudo apt-get update -qq
-sudo apt-get install -qq cmake libboost-filesystem-dev libboost-system-dev libboost-serialization-dev libeigen3-dev doxygen graphviz python3-numpy python3-tk python3-matplotlib texlive texlive-latex-extra
+sudo apt-get install -qq cmake libboost-filesystem-dev libboost-system-dev libboost-serialization-dev libeigen3-dev doxygen graphviz python3 python3-numpy python3-tk python3-matplotlib texlive texlive-latex-extra
