### 2019-03-16 5c3bf66
Restructured INSTALL.md
