+++ INSTALL.md
@@ -1,4 +1,16 @@
-#C++: INSTALLING REQUIRED/OPTIONAL PACKAGES FOR DEBIAN (UBUNTU 16.04) USERS
+# Python version of dmpbbo
+
+## Installing required/optional packages for debian users (tested under Ubuntu 16.04) 
+
+dmpbbo requires python3 (mainly due to the way packages are imported), numpy (http://www.numpy.org/) and matplotlib (https://matplotlib.org/), plus some latex packages for the plotting.
+
+`sudo apt-get install python3-numpy python3-tk python3-matplotlib texlive texlive-latex-extra`
+
+# C++ version of dmpbbo
+
+## Installing required/optional packages for debian users (tested under Ubuntu 16.04) 
+
+### Automatic installation
 
 To automatically install the packages (required and optional) run:
 
@@ -6,21 +18,15 @@ To automatically install the packages (required and optional) run:
 
 Otherwise, for manual installation see [below](#manual_installation).
 
-###Required packages:
+### Required packages
 
 `sudo apt-get install cmake libboost-filesystem-dev libboost-system-dev libboost-serialization-dev libeigen3-dev `
 
-###Optional packages:
+### Optional packages
 
 `sudo apt-get install doxygen graphviz`
 
-###Python required packages:
-
-dmpbbo requires python3 (mainly due to the way packages are imported), numpy (http://www.numpy.org/) and matplotlib (https://matplotlib.org/), plus some latex packages for the plotting.
-
-`sudo apt-get install python3-numpy python3-tk python3-matplotlib texlive texlive-latex-extra`
-
-#C++: BUILDING DMPBBO 
+## Building dmpbbo 
 
 Please make sure you have installed all the required packages before trying to compile. 
 We assume that you use a compiler that is compatible with the C++11 standard.
@@ -28,15 +34,12 @@ We assume that you use a compiler that is compatible with the C++11 standard.
 To compile, do:
  `mkdir -p build_dir; cd build_dir; cmake .. -DCMAKE_BUILD_TYPE=Release; make; make Docs`
 
-To compile a version that has debugging symbols included, and compiles some tests also:
-`mkdir -p build_dir_debug; cd build_dir_debug; cmake .. -DCMAKE_BUILD_TYPE=Debug; make VERBOSE=1`
-
 When running the code on a real robot, you want to do an optimized build without asserts, i.e. `-DCMAKE_BUILD_TYPE=Release`
 
 To install libs in `/usr/local/lib`, headers in `/usr/local/include` and binaries in `dmpbbo/bin`:
  `sudo make install`
 
-#C++: STATIC VS SHARED LIBRARIES
+### Static vs. shared libraries
 
 The default is to build static libraries (e.g. `libdmp.a`) and to statically link these libraries to the demo and test executables (the reason for this is the boost bug below).
 
@@ -58,7 +61,14 @@ This is a bug in `boost::serialization` for versions >=1.44, which can be fixed 
 
 Note that the crash happens towards the very end of the program, just before returning from main() when desctructors are called. So it does not affect functionality, but it is just very irritating that boost crashes and sometimes dumps a core (I use boost version 1.48, maybe the issue is fixed in later versions). If I had known things like this would happen, I would have probably used cereal (http://uscilab.github.io/cereal/) rather than `boost::serialization`... 
 
-#C++: REAL-TIME DEBUGGING BUILD
+## Building dmpbbo (advanced build options for debugging purposes)
+
+# Including debugging symbols and tests
+
+To compile a version that has debugging symbols included, and compiles some tests also:
[…]
