### 2019-03-16 a876e51
Update INSTALL.md
