### 2019-03-16 e37daeb
Documented: test executables are not supported.
closes #40
closes #39
