### 2019-03-16 stulp/dmpbbo#29 @stulp
We've rewritten INSTALL.md. If it is satisfactory now, could you close this issue?
