### 2019-03-16 8f1b32c
Install in source dir.
