### 2019-03-16 5faf9ff
Started moving documentation of demos from doxygen to md files
