+++ demos/demos.md
@@ -1,92 +1,29 @@
-This page provides a list of all the demos, in the recommended order. 
+# Demos
 
-\section sec_demos_python_cpp Python/C++
+This directory contains demos to show the functionality of DmpBbo. The cpp files are intented as a basis for your own tests/demos/projects.
 
-The demos for the C++ code are in the demos/ directory, and for Python in the demos_python/ directory. For most of the C++ demos (e.g. demoExponentialSystem.cpp) there is a corresponding Python file (e.g. demoExponentialSystemWrapper.py), which essentially calls a binary (e.g. bin/demoExponentialSystem), and plots the results. (yes, I know about Python bindings; historical reasons)
+Many of the compiled executables are accompanied by a Python wrapper, which calls the executable, and reads the files it writes, and then plots them (yes, I know about Python bindings; this approach allows better debugging of the format of the output files, which should always remain compatible between the C++ and Python versions of DmpBbo). 
 
-Note that there are also test (e.g. ./src/dmp/dynamicalsystems/testDynamicalSystemSerialization.cpp), which are intended for debugging purposes, not to understand the code. These test are not well-documented, not included in the Doxygen documentation.
+If you are not interested in C++, the pure Python demos are located in dmpbbo/demos_python/.
 
-\section sec_demos_list  List of demos
-
-Here is a list of the available demos, in the recommended order 
-
-\ref FunctionApproximators
-
-\li demoLeastSquares.cpp: Demonstrates how to run least squares regression
-\li demoFunctionApproximatorTraining.cpp: Demonstrates how to train all functionapproximators with 1D and 2D data.
-
-\ref DynamicalSystems
-
-\li demoExponentialSystem.cpp:  Demonstrates how to initialize and integrate an exponential dynamical system.
-\li demoDynamicalSystems.cpp:  Demonstrates how to initialize, integrate, perturb all implemented exponential systems.
-
-\ref Dmps
-
-\li demoDmp.cpp:  Demonstrates how to initialize, train and integrate a Dmp.
-\li demoDmpTrainFromTrajectoryFile.cpp:  Demonstrates how to train a Dmp with a trajectory in a txt file.
-\li demoDmpChangeGoal.cpp:  Demonstrates how to change the goal for a Dmp, and the effects of different scaling approaches.
-
-\li demoDmpContextual.cpp:  Demonstrates how to initialize, train and integrate a Contextual Dmp.
-\li demoDmpContextualGoal.cpp:  Demonstrates how to initialize, train and integrate a Contextual Dmp that adapts the goal state to the task parameters.
+The bottom-up approach to understanding the functionality of the code would imply the following order:
 
 
-\ref BBO
+* Dynamical Systems (in dynamicalsystems/) This module provides implementations of several basic dynamical systems. DMPs are combinations of such systems. This module is completely independent of all other modules.
 
-\li demoOptimization.cpp:  Demonstrates how to run an evolution strategy to optimize a distance function, implemented as a CostFunction.
+* Function Approximation (in functionapproximators/) This module provides implementations (but mostly wrappers around external libraries) of several function approximators. DMPs use function approximators to learn and reproduce arbitrary smooth movements. This module is completely independent of all other modules.
 
+* Dynamical Movement Primitives (in dmp/) This module provides an implementation of several types of DMPs. It depends on both the DynamicalSystems and FunctionApproximators modules, but no other.
 
-\ref DMP_BBO
[…]
