+++ demos/dmp/demos.md
@@ -1,0 +1,8 @@
+\ref Dmps
+
+\li demoDmp.cpp:  Demonstrates how to initialize, train and integrate a Dmp.
+\li demoDmpTrainFromTrajectoryFile.cpp:  Demonstrates how to train a Dmp with a trajectory in a txt file.
+\li demoDmpChangeGoal.cpp:  Demonstrates how to change the goal for a Dmp, and the effects of different scaling approaches.
+
+\li demoDmpContextual.cpp:  Demonstrates how to initialize, train and integrate a Contextual Dmp.
+\li demoDmpContextualGoal.cpp:  Demonstrates how to initialize, train and integrate a Contextual Dmp that adapts the goal state to the task parameters.
