+++ demos/functionapproximators/demos.md
@@ -1,0 +1,5 @@
+\ref FunctionApproximators
+
+\li demoLeastSquares.cpp: Demonstrates how to run least squares regression
+\li demoFunctionApproximatorTraining.cpp: Demonstrates how to train all functionapproximators with 1D and 2D data.
+
