+++ demos/bbo/demos.md
@@ -1,0 +1,3 @@
+\ref BBO
+
+\li demoOptimization.cpp:  Demonstrates how to run an evolution strategy to optimize a distance function, implemented as a CostFunction.
