+++ demos/dmp_bbo/demos.md
@@ -1,0 +1,6 @@
+\ref DMP_BBO
+
+\li demoOptimizationTask.cpp:  Demonstrates how to run an evolution strategy to optimize the parameters of a quadratic function, implemented as a Task and TaskSolver.
+\li demoOptimizationDmp.cpp:  Demonstrates how to run an evolution strategy to optimize a Dmp.
+\li demoImitationAndOptimization.cpp:  Demonstrates how to initialize a DMP with a trajectory, and then optimize it with an evolution strategy.
+\li demoOptimizationDmpArm2D.cpp:  Demonstrates how to run an evolution strategy to optimize a Dmp, on a task with a viapoint task with a N-DOF arm in a 2D space.
