### 2019-03-16 ae85807
functionapproximators demos
