+++ demos/functionapproximators/demos.md
@@ -1,5 +1,14 @@
-\ref FunctionApproximators
+# FunctionApproximators Demos
 
-\li demoLeastSquares.cpp: Demonstrates how to run least squares regression
-\li demoFunctionApproximatorTraining.cpp: Demonstrates how to train all functionapproximators with 1D and 2D data.
+## demoLeastSquares.cpp/demoLeastSquaresWrapper.py
+
+Demonstrates how to run least squares regression.
+
+Reads the files inputs.txt, targets.txt and weights.txt from a directory, and runs (weighted) least squares.
+
+Parameters to the executable and Python script are the regularization constant, whether the line can have an intercept with the y-axis, and the directory to read the files from. Call the Python script with --help for more information.
+
+## demoFunctionApproximatorTraining.cpp/demoFunctionApproximatorTrainingWrapper.py
+
+Demonstrates how to train all functionapproximators with 1D and 2D data. The Python wrapper additionaly plots the results.
 
