+++ demos/dmp_bbo/demos.md
@@ -1,6 +1,22 @@
-\ref DMP_BBO
+# Dmp+Bbo demos
 
-\li demoOptimizationTask.cpp:  Demonstrates how to run an evolution strategy to optimize the parameters of a quadratic function, implemented as a Task and TaskSolver.
-\li demoOptimizationDmp.cpp:  Demonstrates how to run an evolution strategy to optimize a Dmp.
-\li demoImitationAndOptimization.cpp:  Demonstrates how to initialize a DMP with a trajectory, and then optimize it with an evolution strategy.
-\li demoOptimizationDmpArm2D.cpp:  Demonstrates how to run an evolution strategy to optimize a Dmp, on a task with a viapoint task with a N-DOF arm in a 2D space.
+## demoOptimizationTask.cpp / demoOptimizationDmpWrapper.py
+
+Demonstrates how to run an evolution strategy to optimize the parameters of a quadratic function, implemented as a Task and TaskSolver.
+
+## demoOptimizationDmp.cpp / demoOptimizationTaskWrapper.py
+
+Demonstrates how to run an evolution strategy to optimize a Dmp.
+
+## demoImitationAndOptimization.cpp
+
+Demonstrates how to initialize a DMP with a trajectory, and then optimize it with an evolution strategy.
+
+## demoOptimizationDmpArm2D.cpp / demoOptimizationDmpArm2DWrapper.py
+
+Demonstrates how to run an evolution strategy to optimize a Dmp, on a task with a viapoint task with a N-DOF arm in a 2D space.
+
+
+
+
+
