+++ demos/bbo/demos.md
@@ -1,3 +1,5 @@
-\ref BBO
+# BBO Demos
 
-\li demoOptimization.cpp:  Demonstrates how to run an evolution strategy to optimize a distance function, implemented as a CostFunction.
+## demoOptimization.cpp / demoOptimizationWrapper.py  
+
+Demonstrates how to run an evolution strategy to optimize a distance function, implemented as a CostFunction.
