+++ src/mainpage.hpp
@@ -120,7 +120,7 @@ See \ref page_serialization
  * 
  * DmpBbo comes with several demos.
  * 
- * The C++ demos are located in the dmpbbo/demos/ directory. Please see the demos.md file located there.
+ * The C++ demos are located in the dmpbbo/demos/ directory. Please see the README.md files located there.
  *
  * Many of the compiled executables are accompanied by a Python wrapper, which calls the executable,  and reads the files it writes, and then plots them (yes, I know about Python bindings; this approach allows better debugging of the format of the output files, which should always remain compatible between the C++ and Python versions of DmpBbo). For completeness, the pure Python demos are located in dmpbbo/demos_python.
  *
