### 2019-03-16 c2269d1
Finished documenting demos in md files. Fixes #32
