### 2019-03-16 f003e08
Generated plotdata.py scripts use relative directory now.
The mini plotdata.py scripts should use relative paths. If the directory with data is moved, the absolute path would no longer be correct.

This means the plotdata.py script should only be called in that directory. Previous documentation stating otherwise has already been removed.

Fixes #35.
