### 2019-03-16 d986db0
Install demo executables in demos/ directory.
Installing them there might seem counterintuivtive. Please see the discussion in #33 for the rationale behind this.

Fixes #33.
