### 2019-03-16 e7b5814
directory is input argument to executable now
