### 2019-03-16 5ccb9d0
Store data locally when running a demo.
This makes it much easier to see:
1) That data is written in the first place
2) What that data is
