### 2019-03-16 #1225 @stulp
Here is a small update. Based on the feedback from the reviewers, following larger modifications have been made (apart from smaller bugfixes). These modifications are related to the demos and tests in the C++ version, not the core library.

- The executables and Python wrappers were in different directories which was confusing. They are now installed in the demos/ directory so that they are together.

- Demos are now documented in README.md files in the subdirectories of the demos/ directory. Again, this is to have everything in one place, and allow users to understand and run demos without calling doxygen (doxygen documentation is now only for the API of the libraries in the src/ directory)

- When running the demos, the data is no longer stored in /tmp, but in the demos/ directory itself. This ensures that everything is in one place, makes it more transparent that data is being written, and allows the user to have a look at the generated data on the spot.

- Tests in dmpbbo are not unit test but programs used for development. They are no longer installed in bin_test to avoid confusion about this.

- INSTALL.txt has been rewritten and improved into INSTALL.md
