### 2019-03-19 #1225 @jenskober
@stulp Thanks for all the updates, looks great now from my side.

@graiola One last thing about the paper itself: if you check the article proof above, on the second page in applications you have "(Stulp,2012,Stulp et al. (2013),stulp14simultaneous,Stulp & Sigaud (2015))" "stulp14simultaneous" is missing in the references, the first one "Stulp,2012" looks strange as well, and a few spaces after the commas (also first page "(A. J. Ijspeert, Nakanishi, & Schaal, 2002,A. Ijspeert, Nakanishi, Pastor, Hoffmann, & Schaal (2013))") would be nice...
