### 2019-03-19 91139ef
Consistent variable names in Dmp tutorial.
