### 2019-03-20 dceb112
Moving tutorial to Markdown files
