+++ src/dmp_bbo/runOptimizationTask.hpp
@@ -101,6 +101,7 @@ bool saveToDirectory(std::string directory, int i_update, const std::vector<Dist
 
 /** \page page_dmp_bbo Black Box Optimization of Dynamical Movement Primitives
 
+
 This page explains how to apply black-box optimization algorithms to the optimization of dynamical movement primitives.
 
 This page assumes you have read the page on \ref page_bbo and \ref page_dmp. When applying BBO to policy improvement (e.g. optimizing a DMP on a robot), the concept of a "rollout" becomes important. A rollout is the result of executing a policy (e.g. a DMP) with a certain set of policy parameters (e.g. the parameter of the DMP). Although the search space for optimization is in the space of the policy parameters, the costs are rather determined from the rollout.
