+++ src/dmp/Dmp.hpp
@@ -451,183 +451,21 @@ BOOST_CLASS_IMPLEMENTATION(DmpBbo::Dmp,boost::serialization::object_serializable
 #endif // _DMP_H_
 
 
+namespace DmpBbo {
+  
 /** \page page_dmp Dynamical Movement Primitives
 
-This page explains the background and implementation of dynamical movement primitives.
+This page provides and overview of the implementation of DMPs in the \c dmps/ module.
 
-\section sec_dmp_introduction Introduction
+It is assumed you have read about the theory behind DMPs in the tutorial <a href="https://github.com/stulp/dmpbbo/tutorials/dmps.md">tutorials/dmps.md</a>. Note that in the tutorial, we have used the notation \f$[z~y]\f$ for consistency with the DMP literature. In the C++ implementation, the order is rather \f$[y~z]\f$.
 
-The core idea behind dynamical movement primitives (DMPs) is to represent movement primitives as a combination of dynamical systems (please read \ref page_dyn_sys, if you haven't already done so). The state variables of the main dynamical system \f$ [\mathbf{y \dot{y} \ddot{y}} ]\f$ then represent trajectories for controlling, for instance, the 7 joints of a robot arm, or its 3D end-effector position. The attractor state is the end-point or \em goal of the movement.
-
-The key advantage of DMPs is that they inherit the nice properties from linear dynamical systems (guaranteed convergence towards the attractor, robustness to perturbations, independence of time, etc) whilst allowing arbitrary (smooth) motions to be represented by adding a non-linear forcing term. This forcing term is often learned from demonstration, and subsequently improved through reinforcement learning.
-
-DMPs were introduced in \cite ijspeert02movement, but in this section we follow largely the notation and description in \cite ijspeert13dynamical, but at a slower pace.
-
-\em Historical \em remark. Recently, the term "dynamicAL movement primitives" is preferred over "dynamic movement primitives". The newer term makes the relation to dynamicAL systems more clear, and avoids confusion about whether the output of "dynamical movement primitives" is in kinematic or dynamic space (it is usually in kinematic space).
-
-\em Remark. This documentation and code focusses only on discrete movement primitives. For rythmic movement primitives, we refer to \cite ijspeert13dynamical.
-
-
-\section sec_core Basic Point-to-Point Movements: A Critically Damped Spring-Damper System
-
-At the heart of the DMP lies a spring-damper system, as described in \ref dyn_sys_spring_damper. In DMP papers, the notation of the spring-damper system is usually a bit different:
-\f{eqnarray*}{
-m\ddot{y}  =& -ky -c\dot{y} & \mbox{spring-damper system, ``traditional notation''} \\
-m\ddot{y}  =& c(-\frac{k}{c}y - \dot{y})\\
-\tau\ddot{y}  =& \alpha(-\beta y - \dot{y})      & \mbox{with } \alpha=c,~~\beta = \frac{k}{c},~~m=\tau\\
-\tau\ddot{y}  =& \alpha(-\beta (y-y^g) - \dot{y})& \mbox{with attractor } y^g\\
-\tau\ddot{y}  =& \alpha(\beta (y^g-y) - \dot{y})& \mbox{typical DMP notation for spring-damper system}\\
-\f}
-
-In the last two steps, we change the attractor state from 0 to \f$y^g\f$, where \f$y^g\f$ is the goal of the movement.
-
[…]
