### 2019-03-27 #1225 @studywolf
a few issues submitted to the repo as i was going through the examples. I haven't fully confirmed the functionality or look at the functionality documentation.

There are no unit tests for checking off the 'automated tests' box, I see it was mentioned above but didn't see it specifically addressed. What's the resolution on this?
