### 2019-03-27 113b10a
demoDynamicalSystems tests for valid user entries now.
Fixes #43
