### 2019-03-27 f18090a
Plot end and start of trajectory.
Fixes #42
