### 2019-03-27 a0072a2
Removed implementation aspects from tutorial
