### 2019-03-27 bcc9c71
Moving tutorials from doxygen to markdown files
