+++ src/dmp_bbo/runOptimizationTask.hpp
@@ -101,183 +101,7 @@ bool saveToDirectory(std::string directory, int i_update, const std::vector<Dist
 
 /** \page page_dmp_bbo Black Box Optimization of Dynamical Movement Primitives
 
-
-This page explains how to apply black-box optimization algorithms to the optimization of dynamical movement primitives.
-
-This page assumes you have read the page on \ref page_bbo and \ref page_dmp. When applying BBO to policy improvement (e.g. optimizing a DMP on a robot), the concept of a "rollout" becomes important. A rollout is the result of executing a policy (e.g. a DMP) with a certain set of policy parameters (e.g. the parameter of the DMP). Although the search space for optimization is in the space of the policy parameters, the costs are rather determined from the rollout.
-
-From an implementation point of view, applying BBO to policy improvement (which may execute DMPs) requires several extensions:
-
-\li Running multiple optimizations in parallel, one for each DOF of the DMP
-\li Using rollouts with a Task/TaskSolver instead of a CostFunction
-\li Allowing the optimization to be run one update at a time, instead of in a loop
-
-\section sec_parallel_optimization Optimizing DMP DOFs in parallel
-
-Consider a 7-DOF robotic arm. With a DMP, each DOF would be represented by a different dimension of a 7-D DMP (let as assume that each dimension of the DMP uses 10 basis functions). There are then two distinct approaches towards optimizing the parameters of the DMP with respect to a cost function:
-
-\li Consider the search space for optimization to be 70-D, i.e. consider all open parameters of the DMP in one search space. For this you can use runOptimizationTask(), see also the demos in demoOptimizationTask.cpp and demoOptimizationDmp.cpp. 
-The disadvantage is that reward-weighted averaging will require more samples to perform a robust update, especially for the covariance matrix, which is 70X70! 
-
-\li Run a seperate optimization for each of the 7 DOFs. Thus, here we have seven 10-D search spaces (with different distributions and samples for each DOF), but using the same costs for all optimizations. This is what is proposed for instance in PI^2. This approach is not able to exploit covariance between different DOFs, but in practice it works robustly. For this you must runOptimizationParallel(), see also the demo in demoOptimizationDmpParallel.cpp. The main difference is that instead of using one distribution (mean and covariance), D distributions must be used (these are stored in a std::vector), one for each DOF.
-
-
-\section sec_bbo_task_and_task_solver CostFunction vs Task/TaskSolver
-
-When the cost function has a simple structure, e.g. cost = \f$ x^2 \f$ it is convenient to implement the function \f$ x^2 \f$ in CostFunction::evaluate(). 
-
-In robotics however, it is suitable to make the distinction between a task (e.g. lift an object), and an entity that solves this task (e.g. your robot, my robot, a simulated robot, etc.). For these cases, the CostFunction is split into a Task and a TaskSolver, as follows:
-
-\code
-CostFunction::evaluate(samples,costs) {
-  TaskSolver::performRollout(samples,cost_vars)
-  Task::evaluateRollout(cost_vars,costs)
-}
-\endcode
-
-The roles of the Task/TaskSolver are:
-
-\li TaskSolver: performing the rollouts on the robot (based on the samples in policy space), and saving all the variables related to computing the costs into a file (called cost_vars.txt)
-
-\li Task: determining the costs from the cost-relevant variables
-
-
-The idea here is that the TaskSolver uses the samples to perform a rollout (e.g. the samples represent the parameters of a policy which is executed) and computes all the variables that are relevant to determining the cost (e.g. it records the forces at the robot's end-effector, if this is something that needs to be minimized)
-
-Some further advantages of this approach:
[…]
