### 2019-03-27 ec25615
Moved images to tutorials (no longer used in doxygen documentation)
