+++ docs/Doxyfile.in
@@ -690,7 +690,7 @@ EXAMPLE_RECURSIVE      = NO
 # directories that contain image that are included in the documentation (see
 # the \image command).
 
-IMAGE_PATH             = @PROJECT_SOURCE_DIR@/docs/images
+IMAGE_PATH             = #@PROJECT_SOURCE_DIR@/tutorials/images
 
 # The INPUT_FILTER tag can be used to specify a program that doxygen should
 # invoke to filter for each input file. Doxygen will invoke the filter program
