### 2019-03-27 7173490
Removed latex tutorial
