+++ docs/tutorial/tutorial.tex
@@ -1,66 +1,0 @@
-\documentclass[a4paper,10pt]{book}
-\usepackage{doxygen_custom}
-\usepackage{enumitem}
-\setlist[1]{itemsep=-5pt,topsep=0pt}
-
-\usepackage{titlesec}
-
-%\titleformat{command}[shape]{format}{label}{sep}{before}[after]
-%\titleformat{\section}      [runin]{\normalfont\LARGE\bfseries}{\thesection}{}{}
-%\titleformat{\subsection}   [runin]{\normalfont\Large\bfseries}{\thesubsection}{}{}
-\titleformat{\subsubsection}[runin]{\normalfont\bfseries}{\thesubsubsection}{1em}{}[.]
-
-
-%\newcommand{\chaptermark}{}
-\input{preamble}
-
-\graphicspath{{../images/}}
-
-\author{Freek Stulp}
-\title{Dynamical Movement Primitives Tutorial}
-
-\begin{document}
-
-\maketitle
-
-This document was generated by doxygen, so the formatting is not always optimized for pdflatex.
-
-\setcounter{tocdepth}{2}
-\tableofcontents
-
-\newpage
-
-
-
-%\input{../latex/index}
-%
-%\newpage
-
-\chapter{Dynamical Systems}
-\input{ ../../build_dir/docs/latex/page_dyn_sys}
-
-\chapter{Dynamical Movement Primitives}
-
-\input{ ../../build_dir/docs/latex/page_dmp}
-
-\bibliographystyle{apalike}
-\bibliography{../dmp}
-
-
-\chapter{Function Approximation for the Forcing Term}
-
-(Please note that the the tutorial on Dynamical Movement Primitives is now basically over. The rest is more about documentation of the code.)
-
-\input{ ../../build_dir/docs/latex/page_func_approx}
-
-\chapter{Black-box Optimization with Evolution Strategies}
-
-\input{ ../../build_dir/docs/latex/page_bbo}
-
-%\chapter{Black-box Optimization of Dynamical Movement Primitives}
-%
-%\input{ ../../build_dir/docs/latex/page_dmp_bbo}
-
-
-
-\end{document}
