+++ tutorial/dmp.md
@@ -1,6 +1,9 @@
 Dynamical Movement Primitives
 ===============
 
+*It is assumed that you have already read the tutorials on <a href="dynamicalsystems.md">Dynamical Systems</a> and <a href="functionapproximators.md">Function Approximation</a>.*
+
+
 The core idea behind dynamical movement primitives (DMPs) is to represent movement primitives as a combination of dynamical systems (please read <a class="el" href="page_dyn_sys.html">Dynamical Systems</a>, if you haven't already done so). The state variables of the main dynamical system ![alt text](formulae/form_0.png "$ [\mathbf{y~\dot{y}~\ddot{y}} ]$")  then represent trajectories for controlling, for instance, the 7 joints of a robot arm, or its 3D end-effector position. The attractor state is the end-point or *goal* of the movement.
 
 The key advantage of DMPs is that they inherit the nice properties from linear dynamical systems (guaranteed convergence towards the attractor, robustness to perturbations, independence of time, etc) whilst allowing arbitrary (smooth) motions to be represented by adding a non-linear forcing term. This forcing term is often learned from demonstration, and subsequently improved through reinforcement learning.
@@ -174,3 +177,13 @@ The core idea in dynamical movement primitives is to combine dynamical systems, 
 Further enhancements can be made by making the system autonomous (by using the output of a phase system rather than time as an input to the function approximator), or having initial velocities and accelerations of 0 (by using a delayed goal system).
 
 Multi-dimensional DMPs are achieved by using multi-dimensional dynamical systems, and learning one function approximator for each dimension. Synchronization of the different dimensions is ensure by coupling them with only <em>one</em> phase system. 
+
+
+Further reading
+---------------
+
+The next tutorials to go to would be:
+* <a href="dmp_bbo.md">Black-Box Optimizaton of Dynamical Movement Primitives</a> (if you already know about black-box optimization)
+* <a href="bbo.md">Black-Box Optimizaton</a> (if you don't)
+
+
