### 2019-03-27 c232235
Removed formatting from links
