+++ tutorial/dmp.md
@@ -8,11 +8,11 @@ The core idea behind dynamical movement primitives (DMPs) is to represent moveme
 
 The key advantage of DMPs is that they inherit the nice properties from linear dynamical systems (guaranteed convergence towards the attractor, robustness to perturbations, independence of time, etc) whilst allowing arbitrary (smooth) motions to be represented by adding a non-linear forcing term. This forcing term is often learned from demonstration, and subsequently improved through reinforcement learning.
 
-DMPs were introduced in <a class="el" href="citelist.html#CITEREF_ijspeert02movement">[3]</a>, but in this section we follow largely the notation and description in <a class="el" href="citelist.html#CITEREF_ijspeert13dynamical">[4]</a>, but at a slower pace.
+DMPs were introduced in <a href="citelist.html#CITEREF_ijspeert02movement">[3]</a>, but in this section we follow largely the notation and description in <a href="citelist.html#CITEREF_ijspeert13dynamical">[4]</a>, but at a slower pace.
 
 *Historical remark*. The term "dynamicAL movement primitives" is now preferred over "dynamic movement primitives". The newer term makes the relation to dynamicAL systems more clear, and avoids confusion about whether the output of "dynamical movement primitives" is in kinematic or dynamic space (it is usually in kinematic space).
 
-*Remark*. This documentation and code focusses only on discrete movement primitives. For rythmic movement primitives, we refer to <a class="el" href="citelist.html#CITEREF_ijspeert13dynamical">[4]</a>.
+*Remark*. This documentation and code focusses only on discrete movement primitives. For rythmic movement primitives, we refer to <a href="citelist.html#CITEREF_ijspeert13dynamical">[4]</a>.
 
 
 <a name="sec_core"></a>
@@ -28,9 +28,9 @@ At the heart of the DMP lies a spring-damper system, as described in <a href="dy
 
 In the last two steps, we change the attractor state from 0 to ![alt text](formulae/form_2.png "$y^g$") , where ![alt text](formulae/form_2.png "$y^g$")  is the goal of the movement.
 
-To avoid overshooting or slow convergence towards ![alt text](formulae/form_2.png "$y^g$") , we prefer to have a <em>critically</em> <em>damped</em> spring-damper system for the DMP. For such systems ![alt text](formulae/form_3.png "$c = 2\sqrt{mk}$")  must hold, see <a class="el" href="dynamicalsystems.md#dyn_sys_critical_damping">Critical Damping</a>. In our notation this becomes ![alt text](formulae/form_4.png "$\alpha = 2\sqrt{\alpha\beta}$") , which leads to ![alt text](formulae/form_5.png "$\beta = \alpha/4$") . This determines the value of ![alt text](formulae/form_6.png "$\beta$")  for a given value of ![alt text](formulae/form_7.png "$\alpha$")  in DMPs. The influence of ![alt text](formulae/form_7.png "$\alpha$")  is illustrated in the first figure <a class="el" href="dynamicalsystems.md">here</a>.
[…]
