+++ tutorial/dmp_bbo.md
@@ -17,7 +17,7 @@ From an implementation point of view, applying BBO to policy improvement (which 
 
 Consider a 7-DOF robotic arm. With a DMP, each DOF would be represented by a different dimension of a 7-D DMP (let as assume that each dimension of the DMP uses 10 basis functions). There are then two distinct approaches towards optimizing the parameters of the DMP with respect to a cost function:
 
-* Consider the search space for optimization to be 70-D, i.e. consider all open parameters of the DMP in one search space. For this you can use <a class="el" href="runOptimizationTask_8cpp.html#ac0e202635bf044eaa92706e19d340d41" title="Run an evolutionary optimization process, see Black Box Optimization. ">runOptimizationTask()</a>, see also the demos in demoOptimizationTask.cpp and demoOptimizationDmp.cpp. The disadvantage is that reward-weighted averaging will require more samples to perform a robust update, especially for the covariance matrix, which is 70X70!
+* Consider the search space for optimization to be 70-D, i.e. consider all open parameters of the DMP in one search space. For this you can use <a href="runOptimizationTask_8cpp.html#ac0e202635bf044eaa92706e19d340d41" title="Run an evolutionary optimization process, see Black Box Optimization. ">runOptimizationTask()</a>, see also the demos in demoOptimizationTask.cpp and demoOptimizationDmp.cpp. The disadvantage is that reward-weighted averaging will require more samples to perform a robust update, especially for the covariance matrix, which is 70X70!
 * Run a seperate optimization for each of the 7 DOFs. Thus, here we have seven 10-D search spaces (with different distributions and samples for each DOF), but using the same costs for all optimizations. This is what is proposed for instance in PI^2. This approach is not able to exploit covariance between different DOFs, but in practice it works robustly. For this you must runOptimizationParallel(), see also the demo in demoOptimizationDmpParallel.cpp. The main difference is that instead of using one distribution (mean and covariance), D distributions must be used (these are stored in a std::vector), one for each DOF.
 
 
