### 2019-03-27 da03fdc
Bash syntax highlighting
