### 2019-03-27 6fc9e05
Fixed links to tutorials
