+++ src/dmp_bbo/runOptimizationTask.hpp
@@ -101,7 +101,7 @@ bool saveToDirectory(std::string directory, int i_update, const std::vector<Dist
 
 /** \page page_dmp_bbo Black Box Optimization of Dynamical Movement Primitives
 
-The documentation for this module is in the tutorial <a href="https://github.com/stulp/dmpbbo/tutorials/dmp_bbo.md">tutorials/dmp_bbo.md</a>.
+The documentation for this module is in the tutorial <a href="https://github.com/stulp/dmpbbo/tutorial/dmp_bbo.md">tutorial/dmp_bbo.md</a>.
 
  */
 
