+++ src/functionapproximators/FunctionApproximator.hpp
@@ -367,7 +367,7 @@ namespace DmpBbo {
 
 This page provides and overview of the implementation of function approximators in the \c dynamicalsystems/ module.
 
-It is assumed you have read about the theory behind function approximators in the tutorial <a href="https://github.com/stulp/dmpbbo/tutorials/functionapproximators.md">tutorials/functionapproximators.md</a>.
+It is assumed you have read about the theory behind function approximators in the tutorial <a href="https://github.com/stulp/dmpbbo/tutorial/functionapproximators.md">tutorial/functionapproximators.md</a>.
 
 \section sec_fa Function Approximation
 
