### 2019-03-27 d6df42a
Added tex and pdf of tutorial for backwards compatibility.
