+++ docs/tutorial/tutorial.tex
@@ -1,0 +1,6 @@
+\documentclass{article}
+\usepackage{hyperref}
+
+\begin{document}
+This tutorial has moved to \url{http://github.com/stulp/dmpbbo/tutorial}
+\end{document}
