+++ demos/dmp_bbo_robot/README.md
@@ -1,4 +1,3 @@
 # Demos and tips for running dmpbbo on a real robot
 
-This is demo is documented in the doxygen documentation, in the page "Black Box Optimization of Dynamical Movement Primitives". Please compile the doxygen documentation to read it (see INSTALL.md).
-
+Please see the tutorial on this: http://github.com/stulp/dmpbbo/tutorial/dmp_bbo_robot.md
