+++ README.md
@@ -11,7 +11,7 @@ of dynamical movement primitives.
 
 This library may be useful for you if you
 
-+ are interested in the theory behind dynamical movement primitives and their optimization. Then the <a href="https://github.com/stulp/dmpbbo/tutorials"><b>tutorials</b></a> are the best place to start.
++ are interested in the theory behind dynamical movement primitives and their optimization. Then the <a href="https://github.com/stulp/dmpbbo/tutorial"><b>tutorials</b></a> are the best place to start.
 
 + already know about dynamical movement primitives and reinforcement learning, but would rather use existing, tested code than brew it yourself. In this case, the <a href="https://github.com/stulp/dmpbbo/demos"><b>demos</b></a> are a good starting point, as they provide examples of how to use the code.
 
