+++ tutorial/README.md
@@ -1,0 +1,5 @@
+# Tutorials
+
+This directory contains a set of tutorials to understand the black-box optimization of dynamical movement primitives. 
+
+Jump to any tutorial you like. The bottom-up approach would be to start with <a href="tutorial/dynamicalsystems.md">Dynamical Systems</a>.
