+++ tutorial/README.md
@@ -2,4 +2,4 @@
 
 This directory contains a set of tutorials to understand the black-box optimization of dynamical movement primitives. 
 
-Jump to any tutorial you like. The bottom-up approach would be to start with <a href="tutorial/dynamicalsystems.md">Dynamical Systems</a>.
+Jump to any tutorial you like. The bottom-up approach would be to start with <a href="dynamicalsystems.md">Dynamical Systems</a>.
