### 2019-03-27 #1225 @danielskatz
> There are no unit tests for checking off the 'automated tests' box, I see it was mentioned above but didn't see it specifically addressed. What's the resolution on this?

The test do not need need to be automated, but there do need to be some tests.  If there are tests, please check it off.  If there are no tests at all, some need to be added
