### 2019-03-27 #1225 @danielskatz
I would say minimal functionality tests - something that verifies the basic functionality
