### 2019-04-09 #1225 @studywolf
Was able to check off a few more, mostly available time is blocking me. I had one remaining issue on the repo (and just raised another), should I not be waiting for them to be resolved before checking off the corresponding box?
