### 2019-04-13 ca99c98
How to contribute to dmpbbo
Fixes #46.
