### 2019-04-13 d96f8c1
Mixed C++/Python demo in dmp_bbo_robot; not done.
