+++ demos/dmp_bbo_robot/performDmpRollouts.cpp
@@ -1,0 +1,164 @@
+/**
+ * \file trainDmpTrainFromTrajectoryFile.cpp
+ * \author Freek Stulp
+ * \brief  Demonstrates how to train a Dmp with a trajectory in a txt file.
+ *
+ * \ingroup Demos
+ * \ingroup Dmps // zzz
+ *
+ * This file is part of DmpBbo, a set of libraries and programs for the 
+ * black-box optimization of dynamical movement primitives.
+ * Copyright (C) 2014 Freek Stulp, ENSTA-ParisTech
+ * 
+ * DmpBbo is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 2 of the License, or
+ * (at your option) any later version.
+ * 
+ * DmpBbo is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with DmpBbo.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "dmp/Dmp.hpp"
+#include "dmp/Trajectory.hpp"
+
+#include "dynamicalsystems/DynamicalSystem.hpp"
+#include "dynamicalsystems/ExponentialSystem.hpp"
+#include "dynamicalsystems/SigmoidSystem.hpp"
+#include "dynamicalsystems/TimeSystem.hpp"
+#include "dynamicalsystems/SpringDamperSystem.hpp"
+
+#include "functionapproximators/FunctionApproximatorRBFN.hpp"
+#include "functionapproximators/MetaParametersRBFN.hpp"
+#include "functionapproximators/ModelParametersRBFN.hpp"
+#include "functionapproximators/FunctionApproximatorLWR.hpp"
+#include "functionapproximators/MetaParametersLWR.hpp"
+#include "functionapproximators/ModelParametersLWR.hpp"
+
+#include "dmpbbo_io/EigenFileIO.hpp"
+
+#include <boost/filesystem.hpp>
+#include <iostream>
+#include <fstream>
+
+
+using namespace std;
+using namespace Eigen;
+using namespace DmpBbo;
+
+void help(char* binary_name)
+{
+  cout << "Usage: " << binary_name << " <dmp filename.xml> <input/output directory> " << endl;
+}
+
+/** Main function
+ * \param[in] n_args Number of arguments
+ * \param[in] args Arguments themselves
+ * \return Success of exection. 0 if successful.
+ */
+int main(int n_args, char** args)
+{
+  
+  if (n_args!=3)
+  {
+    help(args[0]);
+    return -1;
+  }
+  
+  if (string(args[1]).compare("--help")==0)
+  {
+    help(args[0]);
+    return 0;
+  }
+
+  string dmp_filename = string(args[1]);
+  string directory = string(args[2]);
+  
+  cout << directory << endl;
+  MatrixXd samples;
+  if (!loadMatrix(directory+"samples.txt", samples)) return -1;
+  cout << "  samples=" << samples << endl;
+  
+  std::ifstream ifs(dmp_filename);
+  boost::archive::xml_iarchive ia(ifs);
+  Dmp* dmp;
+  ia >> BOOST_SERIALIZATION_NVP(dmp);
+  ifs.close();
+  
+  cout << *dmp << endl;
+  
+  /*
+  cout << "Reading trajectory from TXT file: " << input_txt_file << endl;
+  Trajectory trajectory = Trajectory::readFromFile(input_txt_file);
+  if (trajectory.length()==0)
+  {
[…]
