+++ demos/CMakeLists.txt
@@ -3,5 +3,6 @@ add_subdirectory(dynamicalsystems)
 add_subdirectory(dmp)
 add_subdirectory(bbo)
 add_subdirectory(dmp_bbo)
+add_subdirectory(dmp_bbo_robot)
 
 
