### 2019-04-14 #1225 @studywolf
@danielskatz i've completed the review, and with the contributing documentation added it looks good aside from the documentation for one part of the code. This is currently an open issue on the github, so deferring to you on this. If you think it's fine to not have this resolved before approving then :+1:
