### 2019-04-14 #1225 @danielskatz
👋 @graiola - what is the current version number?

And please make sure there is an archive of the current version, and let me know the DOI for it as well.
