### 2019-04-15 #1225 @stulp
It would be nice if I could close issue 44 (which was opened by @studywolf) before assigning the next version number. That is the version I would like the DOI to be assigned with.

Fixing 44 is quite a bit of work (involving an overhaul of the serialization functions), but I should be done within a week or so. Would it be possible to wait with the acceptance until I have fixed it?
