### 2019-04-15 #1225 @danielskatz
If it's that close to done, then yes, let's wait.  My suggestion to go ahead was based on an assumption it would take longer.
