+++ src/functionapproximators/ModelParametersRBFN.cpp
@@ -20,22 +20,12 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with DmpBbo.  If not, see <http://www.gnu.org/licenses/>.
  */
- 
-#include <boost/serialization/export.hpp>
-#include <boost/archive/text_iarchive.hpp>
-#include <boost/archive/text_oarchive.hpp>
-#include <boost/archive/xml_iarchive.hpp>
-#include <boost/archive/xml_oarchive.hpp>
-#include "functionapproximators/FunctionApproximator.hpp"
+
 #include "functionapproximators/ModelParametersRBFN.hpp"
-
-/** For boost::serialization. See http://www.boost.org/doc/libs/1_55_0/libs/serialization/doc/special.html#export */
-BOOST_CLASS_EXPORT_IMPLEMENT(DmpBbo::ModelParametersRBFN);
-
-#include "dmpbbo_io/BoostSerializationToString.hpp"
-#include "dmpbbo_io/EigenBoostSerialization.hpp"
 #include "functionapproximators/BasisFunction.hpp"
 #include "functionapproximators/UnifiedModel.hpp"
+
+#include "dmpbbo_io/BoostSerializationToString.hpp"
 
 #include <iostream>
 #include <fstream>
@@ -110,19 +100,6 @@ void ModelParametersRBFN::kernelActivations(const Eigen::Ref<const Eigen::Matrix
     kernel_activations_cached_ = kernel_activations;
   }
   
-}
-
-template<class Archive>
-void ModelParametersRBFN::serialize(Archive & ar, const unsigned int version)
-{
-  // serialize base class information
-  ar & BOOST_SERIALIZATION_BASE_OBJECT_NVP(ModelParameters);
-
-  ar & BOOST_SERIALIZATION_NVP(centers_);
-  ar & BOOST_SERIALIZATION_NVP(widths_);
-  ar & BOOST_SERIALIZATION_NVP(weights_);
-  ar & BOOST_SERIALIZATION_NVP(all_values_vector_size_);
-  ar & BOOST_SERIALIZATION_NVP(caching_);
 }
 
 string ModelParametersRBFN::toString(void) const
