+++ src/functionapproximators/MetaParametersRRRFF.cpp
@@ -21,17 +21,8 @@
  * along with DmpBbo.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <boost/serialization/export.hpp>
-#include <boost/archive/text_iarchive.hpp>
-#include <boost/archive/text_oarchive.hpp>
-#include <boost/archive/xml_iarchive.hpp>
-#include <boost/archive/xml_oarchive.hpp>
 #include "functionapproximators/MetaParametersRRRFF.hpp"
 
-/** For boost::serialization. See http://www.boost.org/doc/libs/1_55_0/libs/serialization/doc/special.html#export */
-BOOST_CLASS_EXPORT_IMPLEMENT(DmpBbo::MetaParametersRRRFF);
-
-#include "dmpbbo_io/EigenBoostSerialization.hpp"
 #include "dmpbbo_io/BoostSerializationToString.hpp"
 
 #include <iostream>
@@ -53,17 +44,6 @@ MetaParametersRRRFF::MetaParametersRRRFF(int expected_input_dim, int number_of_b
 MetaParametersRRRFF* MetaParametersRRRFF::clone(void) const 
 {
   return new MetaParametersRRRFF(getExpectedInputDim(),number_of_basis_functions_,regularization_,gamma_);
-}
-
-template<class Archive>
-void MetaParametersRRRFF::serialize(Archive & ar, const unsigned int version)
-{
-  // serialize base class information
-  ar & BOOST_SERIALIZATION_BASE_OBJECT_NVP(MetaParameters);
-
-  ar & BOOST_SERIALIZATION_NVP(number_of_basis_functions_);
-  ar & BOOST_SERIALIZATION_NVP(regularization_);
-  ar & BOOST_SERIALIZATION_NVP(gamma_);
 }
 
 string MetaParametersRRRFF::toString(void) const 

[…]
