### 2019-04-16 3324a27
Serialization for bbo module
