### 2019-04-16 1432f76
Library compilation works, but linking does not due to opaque boost::serialization errors.
