+++ demos/dmp_bbo/demoOptimizationDmpParallel.cpp
@@ -24,14 +24,14 @@
  * along with DmpBbo.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <string>
-#include <set>
-#include <boost/filesystem.hpp>
-#include <eigen3/Eigen/Core>
-
 #include "dmp_bbo/runOptimizationTask.hpp"
 #include "dmp_bbo/tasks/TaskViapoint.hpp"
 #include "dmp_bbo/TaskSolverDmp.hpp"
+
+
+#include <boost/archive/xml_oarchive.hpp>
+#include <boost/archive/xml_iarchive.hpp>
+#include "dmp_bbo/serialization.hpp"
 
 #include "dmp/Dmp.hpp"
 #include "functionapproximators/FunctionApproximatorLWR.hpp"
@@ -40,6 +40,12 @@
 #include "bbo/DistributionGaussian.hpp"
 #include "bbo/Updater.hpp"
 #include "bbo/updaters/UpdaterCovarDecay.hpp"
+
+#include <string>
+#include <set>
+#include <boost/filesystem.hpp>
+#include <eigen3/Eigen/Core>
+
 
 using namespace std;
 using namespace Eigen;
