+++ src/bbo/DistributionGaussian.hpp
@@ -24,8 +24,6 @@
 #ifndef DISTRIBUTIONGAUSSIAN_H
 #define DISTRIBUTIONGAUSSIAN_H   
 
-#include "dmpbbo_io/EigenBoostSerialization.hpp"
-
 #include <vector>
 #include <sstream>
 #include <algorithm>
@@ -34,15 +32,13 @@
 
 #include <boost/random.hpp>
 
-#include <boost/serialization/access.hpp>
-#include <boost/serialization/nvp.hpp>
-
 namespace DmpBbo {
 
 /** \brief A class for representing a Gaussian distribution.
  *
  * This is mainly a wrapper around boost functionality
  * The reason to make the wrapper is to provide functionality for serialization/deserialization.
+ * 2019: Decided to turn serialization off; too much of a pain to keep it working...
  */
 class DistributionGaussian
 {
@@ -122,22 +118,8 @@ private:
   
   /** Maximum eigen value of the covariance matrix of the distribution. This cached variable avoid recomputing it every time maxEigenValue is called. */
   mutable double max_eigen_value_ = -1.0;
-  
-  /** Give boost serialization access to private members. */  
-  friend class boost::serialization::access;
-  
-  /** Serialize class data members to boost archive. 
-   * \param[in] ar Boost archive
-   * \param[in] version Version of the class
-   * See http://www.boost.org/doc/libs/1_55_0/libs/serialization/doc/tutorial.html#simplecase
-   */
-  template<class Archive>
-  void serialize(Archive & ar, const unsigned int version)
-  {
-    ar & BOOST_SERIALIZATION_NVP(mean_);
-    ar & BOOST_SERIALIZATION_NVP(covar_);
-  }
 
+  
 };
 
 }
