+++ src/bbo/updaters/UpdaterCovarAdaptation.hpp
@@ -57,31 +57,6 @@ private:
   Eigen::VectorXd base_level_diagonal_;
   double relative_lower_bound_;
  
-  /**
-   * Default constructor.
-   * \remarks This default constuctor is required for boost::serialization to work. See \ref sec_boost_serialization_ugliness
-   */
-  UpdaterCovarAdaptation(void) {};
-  
-  /** Give boost serialization access to private members. */  
-  friend class boost::serialization::access;
-  
-  /** Serialize class data members to boost archive. 
-   * \param[in] ar Boost archive
-   * \param[in] version Version of the class
-   * See http://www.boost.org/doc/libs/1_55_0/libs/serialization/doc/tutorial.html#simplecase
-   */
-  template<class Archive>
-  void serialize(Archive & ar, const unsigned int version)
-  {
-    // serialize base class information
-    ar & BOOST_SERIALIZATION_BASE_OBJECT_NVP(UpdaterMean);
-    
-    ar & BOOST_SERIALIZATION_NVP(diag_only_);
-    ar & BOOST_SERIALIZATION_NVP(learning_rate_);
-    ar & BOOST_SERIALIZATION_NVP(base_level_diagonal_);
-  }
-
 };
 
 }
