+++ src/bbo/updaters/UpdaterMean.hpp
@@ -69,32 +69,6 @@ protected:
   /** Weighting method */
   std::string weighting_method_;
 
-protected:
-  /**
-   * Default constructor.
-   * \remarks This default constuctor is required for boost::serialization to work. See \ref sec_boost_serialization_ugliness
-   */
-  UpdaterMean(void) {};
-  
-private:
-  /** Give boost serialization access to private members. */  
-  friend class boost::serialization::access;
-  
-  /** Serialize class data members to boost archive. 
-   * \param[in] ar Boost archive
-   * \param[in] version Version of the class
-   * See http://www.boost.org/doc/libs/1_55_0/libs/serialization/doc/tutorial.html#simplecase
-   */
-  template<class Archive>
-  void serialize(Archive & ar, const unsigned int version)
-  {
-    // serialize base class information
-    ar & BOOST_SERIALIZATION_BASE_OBJECT_NVP(Updater);
-    
-    ar & BOOST_SERIALIZATION_NVP(eliteness_);
-    ar & BOOST_SERIALIZATION_NVP(weighting_method_);
-  }
-
 };
 
 }

[…]
