+++ src/dmp_bbo/TaskSolverDmp.hpp
@@ -26,17 +26,9 @@
 
 #include "dmp_bbo/TaskSolver.hpp"
 
-#include "dmp/serialization.hpp"
-
-#include "dmpbbo_io/EigenBoostSerialization.hpp"
-
 #include <string>
 #include <set>
 #include <eigen3/Eigen/Core>
-
-#include <boost/serialization/access.hpp>
-#include <boost/serialization/nvp.hpp>
-
 
 
 namespace DmpBbo {
@@ -82,36 +74,6 @@ public:
    * 
    */
   void set_perturbation(double perturbation_standard_deviation);
-
-protected:
-  /**
-   * Default constructor.
-   * \remarks This default constuctor is required for boost::serialization to work. Since this
-   * constructor should not be called by other classes, it is private (boost::serialization is a
-   * friend)
-   */
-  TaskSolverDmp(void);
-
-private:
-  /** Give boost serialization access to private members. */  
-  friend class boost::serialization::access;
-  
-  /** Serialize class data members to boost archive. 
-   * \param[in] ar Boost archive
-   * \param[in] version Version of the class
-   * See http://www.boost.org/doc/libs/1_55_0/libs/serialization/doc/tutorial.html#simplecase
-   */
-  template<class Archive>
-  void serialize(Archive & ar, const unsigned int version)
-  {
-    // serialize base class information
-    ar & BOOST_SERIALIZATION_BASE_OBJECT_NVP(TaskSolver);
-    
-    ar & BOOST_SERIALIZATION_NVP(dmp_);
-    ar & BOOST_SERIALIZATION_NVP(n_time_steps_);    
-    ar & BOOST_SERIALIZATION_NVP(integrate_time_);
-    ar & BOOST_SERIALIZATION_NVP(use_normalized_parameter_);
-  }
 
 };
 
