### 2019-04-16 1966acb
Made members const (possible now because serialization disabled)
