### 2019-04-16 1a360fa
Small bugfixes to make serialization work.
