### 2019-04-16 c30bbec
Do not compile LWPR when BUILD_SHARED_LIBS is ON
