+++ src/functionapproximators/CMakeLists.txt
@@ -1,7 +1,11 @@
 ################################################################################
 file(GLOB SOURCES leastSquares.cpp BasisFunction.cpp Parameterizable.cpp MetaParameters.cpp ModelParameters.cpp FunctionApproximator.cpp UnifiedModel.cpp *LWR.cpp *RRRFF.cpp *GMR.cpp *RBFN.cpp *GPR.cpp getFunctionApproximatorByName.cpp) 
 if(LWPR_FOUND)
-  file(GLOB SOURCES ${SOURCES} *LWPR.cpp) 
+  if (BUILD_SHARED_LIBS) # LWPR doesn't work with shared libs
+    message(WARNING  "LWPR cannot be used with .so libraries.")
+  else()
+    file(GLOB SOURCES ${SOURCES} *LWPR.cpp) 
+  endif()
 endif()
 
 ################################################################################
@@ -10,7 +14,11 @@ install(TARGETS functionapproximators DESTINATION ${LIB_INSTALL_DIR})
 
 file(GLOB HEADERS *.hpp)
 if(LWPR_FOUND)
-  file(GLOB HEADERS ${HEADERS} *LWPR.hpp) 
+  if (BUILD_SHARED_LIBS) # LWPR doesn't work with shared libs
+    message(WARNING  "LWPR cannot be used with .so libraries.")
+  else()
+    file(GLOB HEADERS ${HEADERS} *LWPR.hpp) 
+  endif()
 endif()
 install(FILES ${HEADERS} DESTINATION ${INCLUDE_INSTALL_DIR}/functionapproximators)
 
