+++ CMakeLists.txt
@@ -27,7 +27,7 @@ set(CMAKE_DISABLE_IN_SOURCE_BUILD ON)
 
 
 # Some flags for CXX
-set(CMAKE_CXX_FLAGS "-Wall -std=c++0x -fPIC -frtti ")
+set(CMAKE_CXX_FLAGS "-Wall -std=c++0x -fPIC")
 set(CMAKE_CXX_FLAGS_DEBUG "-O0 -ggdb")
 set(CMAKE_CXX_FLAGS_RELEASE "-O3 -DNDEBUG")
 
@@ -53,7 +53,11 @@ ENDIF(CMAKE_INSTALL_PREFIX_INITIALIZED_TO_DEFAULT)
 # See if LWPR is installed. 
 find_package(LWPR)
 if(LWPR_FOUND)
-  SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DUSE_LWPR")
+  if (BUILD_SHARED_LIBS) # LWPR doesn't work with shared libs
+    message(WARNING  "LWPR cannot be used with .so libraries.")
+  else()
+    SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DUSE_LWPR")
+  endif()
 endif()
 
 ###############################################################################
