+++ .travis.yml
@@ -8,7 +8,6 @@ compiler:
 branches:
   only:
     - master
-    - serialize_tmp
 
 env:
   - DMP_BUILD_TYPE=Debug BUILD_SHARED_LIBS=ON
