+++ src/dmp/Trajectory.cpp
@@ -263,6 +263,13 @@ bool Trajectory::saveToFile(string directory, string filename, bool overwrite) c
   return saveMatrix(directory, filename, traj_matrix, overwrite);  
 }
 
+bool Trajectory::saveToFile(string filename, bool overwrite) const
+{
+  MatrixXd traj_matrix;
+  asMatrix(traj_matrix);
+  return saveMatrix(filename, traj_matrix, overwrite);  
+}
+
 Trajectory Trajectory::readFromFile(string filename, int n_dims_misc)
 {
   MatrixXd traj_matrix;
