### 2019-04-17 9bf76e3
Implementing different steps for running dmp_bbo on a real robot.
