### 2019-04-18 605576b
Save rollout_eval also
