### 2019-04-18 00110b7
Overhaul of demos/dmp_bbo_robot/
