+++ tutorial/dmp_bbo_robot.md
@@ -1,47 +1,48 @@
-Step-by-step howto for training and optimizing DMPs on a real robot
+Step-by-step howto for training and optimizing a DMP on a real robot
 ===============
 
 *It is assumed that you have already read the tutorials on <a href="dmp_bbo.md">Black Box Optimization of Dynamical Movement Primitives</a>.* 
 
-This tutorial will describe the steps involved in training and optimizing a DMP on a real robot. Probably the easiest way to get dmpbbo running for your robot is to copy this directory cp -uva dmp_bbo_robot my_own_optimization, and adapt the cpp and py files to your robot and task.
+This tutorial will describe the steps involved in training and optimizing a DMP on a real robot. Probably the easiest way to get dmpbbo running for your robot is to copy this directory `cp -uva dmp_bbo_robot my_own_optimization`, and adapt the cpp and py files to your robot and task.
 
 ## Step 1: Train the DMP with a demonstration
 
 It is common practice to initialize the DMP with a demonstrated trajectory, so that the optimization does not have to start from scratch. Given that the optimization algorithms are local, such an initialization is essential to avoid local minima that do not solve the task.
 
-This initialization is done with the following command (the binary is compiled from step1A_trainDmpFromTrajectoryFile.cpp):
+This initialization is done with the following command (the binary is compiled from `step1A_trainDmpFromTrajectoryFile.cpp`):
 
     ./step1A_trainDmpFromTrajectoryFile trajectory.txt results/dmp.xml results/policy_parameters.txt results/train/ 6
 
-This reads a trajectory from a file (for the format see the Trajectory class), trains a DMP, and saves it to an xml file with boost::serialization. The policy parameters of this DMP, which are to be optimized in the subsequent optimization step, are also stored in policy_parameters.txt. To analyze the fitting process and tune training meta-parameters, the intermediate fitting results are written to results/train/. These can be plotted with:
+This reads a trajectory from a file (for the format see the Trajectory class), trains a DMP, and saves it to an xml file with boost::serialization. The policy parameters of this DMP, which are to be optimized in the subsequent optimization step, are also stored in `policy_parameters.txt`. To analyze the fitting process and tune training meta-parameters, the intermediate fitting results are written to `results/train/`. These can be plotted with:
 
      python3 step1B_trainDmpFromTrajectoryFilePlot.py results/train/
 
-The "6" in the call to step1A_trainDmpFromTrajectoryFile is a meta-parameter, in this example the number of basis functions. You can try different values and plot the results until a good fit is achieved.
+The "6" in the call to `step1A_trainDmpFromTrajectoryFile` is a meta-parameter, in this example the number of basis functions. You can try different values and plot the results until a good fit is achieved.
 
 ## Step 2: Define the task (i.e. cost function) and implement executing rollouts on the robot
 
-Defining the task requires you to make a class that inherits from dmp_bbo.Task, and implements the following functions:
-* evaluateRollout(cost_vars,...). This is the cost function, which takes the cost-relevant variables (cost_vars) as an input, and returns the cost associated with the rollout. cost_vars thus defines the variables the robot needs to record when performing a rollout, as these variables are required to compute the cost.
-* plotRollout(cost_vars,...). This function visualizes one rollout.
+Defining the task requires you to make a class that inherits from `dmp_bbo.Task`, and implements the following functions:
+
+* `evaluateRollout(cost_vars,...)`. This is the cost function, which takes the cost-relevant variables (cost_vars) as an input, and returns the cost associated with the rollout. cost_vars thus defines the variables the robot needs to record when performing a rollout, as these variables are required to compute the cost.
[…]
