### 2019-04-18 487672f
Replaced viapoint task with a throwing the ball task.
