### 2019-04-19 858ddfe
Better plotting
