+++ tutorial/dmp_bbo_robot.md
@@ -5,19 +5,43 @@ Step-by-step howto for training and optimizing a DMP on a real robot
 
 This tutorial will describe the steps involved in training and optimizing a DMP on a real robot. Probably the easiest way to get dmpbbo running for your robot is to copy this directory `cp -uva dmp_bbo_robot my_own_optimization`, and adapt the cpp and py files to your robot and task.
 
+In the task considered in this tutorial, the robot has to throw a ball into a certain area, as illustrated below. The "robot" makes an elliptical movement with its end-effector (blue trajectory), releases the ball (black circles) after 0.6 seconds, so that the ball flies through the air (green trajectory) until it hits the ground. The aim is to throw the ball to a particular position (the green marker on the "floor"). There is a margin of error, illustrated by the dent in the floor.
+
+![alt text](images/task_throw_ball.png  "Illustration of the ball throwing task.")
+
+
 ## Step 1: Train the DMP with a demonstration
 
 It is common practice to initialize the DMP with a demonstrated trajectory, so that the optimization does not have to start from scratch. Given that the optimization algorithms are local, such an initialization is essential to avoid local minima that do not solve the task.
 
 This initialization is done with the following command (the binary is compiled from `step1A_trainDmpFromTrajectoryFile.cpp`):
 
-    ./step1A_trainDmpFromTrajectoryFile trajectory.txt results/dmp.xml results/policy_parameters.txt results/train/ 6
+    ./step1A_trainDmpFromTrajectoryFile trajectory.txt results/dmp.xml results/policy_parameters.txt results/train/ 5
 
 This reads a trajectory from a file (for the format see the Trajectory class), trains a DMP, and saves it to an xml file with boost::serialization. The policy parameters of this DMP, which are to be optimized in the subsequent optimization step, are also stored in `policy_parameters.txt`. To analyze the fitting process and tune training meta-parameters, the intermediate fitting results are written to `results/train/`. These can be plotted with:
 
      python3 step1B_trainDmpFromTrajectoryFilePlot.py results/train/
 
-The "6" in the call to `step1A_trainDmpFromTrajectoryFile` is a meta-parameter, in this example the number of basis functions. You can try different values and plot the results until a good fit is achieved.
+The "5" in the call to `step1A_trainDmpFromTrajectoryFile` is a meta-parameter, in this example the number of basis functions. The first plot shows the results of the function approximation inside each DMP dimension, where the black dots are the input data, the blue line is the approximation, and the vertical red lines is the difference between them (the "residuals"). The x and y-axis are annotated with the vague "input" and "output". This is because the function approximator inside the DMP has not clue what it is fitting. 
+
+![alt text](images/step1_function_approximation_5.png  "Results of function approximation with 5 basis functions.")
+
+The third plot shows the demonstrated trajectory and the reproduced trajectory (the second plot shows more detailed results of integrating the trained DMP). The demonstrated trajectory is the eliptical movement illustrated in the figure above.
+
+![alt text](images/step1_trajectory_comparison_5.png  "Reproduced trajectory with 5 basis functions.")
+
+In the results above, the fitting is not very good, i.e. there is a large difference between the input data and the fitted function. We therefore also see that the reproduced trajectory does not fit the demonstrated trajectory well. This is because 5 basis function do not suffice to fit the data accurately. So let's increase the number of basis functions to 10
+
+    ./step1A_trainDmpFromTrajectoryFile trajectory.txt results/dmp.xml results/policy_parameters.txt results/train/ 10
+     python3 step1B_trainDmpFromTrajectoryFilePlot.py results/train/
+
+Which yields the following result:
+
[…]
