### 2019-04-19 aa680ee
Optionally plot tau, and better axis limits
