### 2019-04-19 8a83c87
Plot residuals first to avoid overlap.
