### 2019-04-19 86c358c
Store more information when training a Dmp
