### 2019-04-19 b45ce9e
Improved plotting
