### 2019-04-19 86b5c3a
Improved plotting
