+++ demos/dmp_bbo_robot/step1B_trainDmpFromTrajectoryFilePlot.py
@@ -34,6 +34,13 @@ sys.path.append(lib_path)
 from functionapproximators.functionapproximators_plotting import * 
 from dmp.dmp_plotting import * 
 
+def legendWithoutDuplicates(ax):
+    handles, labels = ax.get_legend_handles_labels()  
+    lgd = dict(zip(labels, handles))
+    ax.legend(lgd.values(), lgd.keys())
+#    plt.legend()
+    
+
 if __name__=='__main__':
     # Call the executable with the directory to which results should be written
     parser = argparse.ArgumentParser()
@@ -42,10 +49,11 @@ if __name__=='__main__':
     
     directory = args.directory
     
-    print("Plotting")
-    
-    fig = plt.figure(1)
-    axs = [ fig.add_subplot(131), fig.add_subplot(132), fig.add_subplot(133) ] 
+    # Plot trajectories    
+    golden_ratio = 1.618 # graphs have nice proportions with this ratio
+    n_subplots = 3
+    fig3 = plt.figure(3,figsize=(golden_ratio*n_subplots*3,3))
+    axs = [ fig3.add_subplot(1,n_subplots,ii+1) for ii in range(n_subplots) ]
     
     traj_demonstration = numpy.loadtxt(directory+"/traj_demonstration.txt")
     lines = plotTrajectory(traj_demonstration,axs)
@@ -55,29 +63,36 @@ if __name__=='__main__':
     lines = plotTrajectory(traj_reproduced,axs)
     plt.setp(lines, linestyle='--', linewidth=2, color=(0.0,0.0,0.5), label='reproduced')
     
+    legendWithoutDuplicates(axs[-1])
+    
+    fig3.canvas.set_window_title('Comparison between demonstration and reproduced') 
+    plt.tight_layout()
+    
+
+    # Plot results of integrating the DMP
+    tau           = numpy.loadtxt(directory+'/tau.txt')
+    ts_xs_xds     = numpy.loadtxt(directory+'/reproduced_ts_xs_xds.txt')
+    forcing_terms = numpy.loadtxt(directory+'/reproduced_forcing_terms.txt')
+    fa_output     = numpy.loadtxt(directory+'/reproduced_fa_output.txt')
+    
+    fig2 = plt.figure(2,figsize=(golden_ratio*5*3,3*3))
+    plotDmp(ts_xs_xds,fig2,forcing_terms,fa_output,[],tau)
+    fig2.canvas.set_window_title('DMP integration') 
+    plt.tight_layout()
+
+
+    # Plot the results of training the function approximation 
     n_dims = (traj_demonstration.shape[1]-1)//3 # -1 for time, /3 for x/xd/xdd 
-    
-    fig = plt.figure(2,figsize=(n_dims*5,5))
+    fig1 = plt.figure(1,figsize=(golden_ratio*n_dims*4,4))
     for i_dim in range(n_dims):
-        ax = fig.add_subplot(1, n_dims, i_dim+1)
+        ax = fig1.add_subplot(1, n_dims, i_dim+1)
         plotFunctionApproximatorTrainingFromDirectory(directory+"dim"+str(i_dim),ax)
-       
-    #plt.legend()
-    #fig.canvas.set_window_title('Comparison between demonstration and reproduced') 
     
-    # Read data
-    #ts_xs_xds     = numpy.loadtxt(directory+'/reproduced_ts_xs_xds.txt')
-    #forcing_terms = numpy.loadtxt(directory+'/reproduced_forcing_terms.txt')
-    #fa_output     = numpy.loadtxt(directory+'/reproduced_fa_output.txt')
+    fig1.canvas.set_window_title('Results of function approximation') 
[…]
