### 2019-04-19 ea15085
Save figure to png file
