### 2019-04-19 b66ac4e
Using CMA instead of decay
