### 2019-04-19 4c1f5f0
UpdaterCovarAdaptation takes max_level as an argument now.
All diagonal elements of the covar matrix are limited so as not to be larger than "max_level"
