+++ tutorial/dmp_bbo_robot.md
@@ -74,27 +74,28 @@ Gathering the information for cost-vars can be non-trivial in practice. For inst
 
 During the stochastic optimization, the parameters of the DMP will be sampled from a Gaussian distribution (which parameters these are is set through the `Parameterizable` class from which `Dmp` inherits. See the "`set<string> parameters_to_optimize`" code in `step1A_trainDmpFromTrajectoryFile.cpp` for an example). The mean of this distribution will be the parameters that resulted from training the DMP with a demonstration through supervised learning. 
 
-The covariance matrix of this distributions determines the magnitude of exploration. It should not be too low, otherwise the stochasticity of the exploration may be smaller than that of the robot movement itself, and no learning can take place. It should also not be too high for safety reasons; your robot may reach acceleration limits, joint limits, or unexpectedly bump into the environment. 
+The covariance matrix of this distributions determines the magnitude of exploration. This magnitude is defined in terms of sigma, where the diagonal of the covariance matrix is initialized with sigma^2. Sigma should not be too low, otherwise the stochasticity of the exploration may be smaller than that of the robot movement itself, and no learning can take place. It should also not be too high for safety reasons; your robot may reach acceleration limits, joint limits, or unexpectedly bump into the environment. 
 
 You can tune this parameter by calling the following three scripts for different exploration magnitudes:
 
-    MAG=0.1      # Exploration magnitude to try (start low!)
+    SIGMA=0.1      # Exploration magnitude to try (start low!)
     N_SAMPLES=10 # Number of samples to generate
     # Generate samples with this magnitude
     # This will save samples to directories 
     #     results/tune_exploration_0.1/rollout001/policy_parameters.txt
     #     results/tune_exploration_0.1/rollout002/policy_parameters.txt
     #     etc
-    python3 step3A_tuneExploration.py results/policy_parameters.txt results/distribution_initial_covar.txt results/tune_exploration_${MAG}/ ${MAG} ${N_SAMPLES}
+    python3 step3A_tuneExploration.py results/policy_parameters.txt results/distribution_initial_covar.txt results/tune_exploration_${SIGMA}/ ${SIGMA} ${N_SAMPLES}
     # Execute the Dmps with sampled parameters on the robot
-    ./step3B_performExplorationRollouts.bash results/dmp.xml results/tune_exploration_${MAG}/
+    ./step3B_performExplorationRollouts.bash results/dmp.xml results/tune_exploration_${SIGMA}/
     # Plot the rollouts to see the variance in the movements
-    python3 step3C_tuneExplorationPlot.py results/tune_exploration_${MAG}/ results/task.p
+    python3 step3C_tuneExplorationPlot.py results/tune_exploration_${SIGMA}/ results/task.p
 
-Below the results of exploring with magnitudes 0.1, 1.0, 10.0, and 100.0. The value 0.1 is probably too low, because there is hardly any variation in the end-effector movement. 100.0 is definitely too high! If you execute this on your robot you are a braver person than I (Quote from the license: "This library is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY". If your bravery breaks your robot, don't blame me!). Given these results, I'd be comfortable with a value between 1.0 and 10.0. We'll continue with 10.0 in this tutorial, as we can't break any robots in simulation.
[…]
