### 2019-04-20 8c08362
Min and max level now specify sigma, not variance.
