### 2019-04-20 38defe0
Tuned cost weight for demo task
