### 2019-04-20 769f7e8
Rename covar_scale to sigma, to make clear it is not variance.
