### 2019-04-20 42eb9e7
Different updaters for demo
