+++ demos/dmp_bbo_robot/step4C_plotOptimization.py
@@ -26,7 +26,7 @@ lib_path = os.path.abspath('../../python')
 sys.path.append(lib_path)
 
 from dmp_bbo.Task import Task
-from dmp_bbo.dmp_bbo_plotting import plotOptimizationRollouts
+from dmp_bbo.dmp_bbo_plotting import *
 
 if __name__=="__main__":
 
@@ -41,6 +41,9 @@ if __name__=="__main__":
     task = pickle.load(open(directory+'/task.p', "rb" ))
     
     # Plot the optimization results (from the files saved to disk)
-    fig = plt.figure(1,figsize=(15, 5))
-    plotOptimizationRollouts(directory,fig,task.plotRollout)
+    golden_ratio = 1.618 # graphs have nice proportions with this ratio
+    n_subplots = 4
+    fig = plt.figure(3,figsize=(golden_ratio*n_subplots*3,3))
+    plotOptimizationRolloutsTask(directory,fig,task)
+    plt.tight_layout()
     plt.show()
