+++ demos/dmp_bbo_robot/todo.txt
@@ -1,21 +1,0 @@
-Move dmp_bbo_robot up one directory
-  
-
-________________________________________________________________________________
-Once merged to master
-
-
-Remove "#install(TARGETS testUpdaters DESTINATION bin_test)"
-
-
-________________________________________________________________________________
-LATER
-
-Make Python scripts robust against missing data, e.g. cost_vars
-
-update Python/C++ comparison
-
-Check if true: An example is given in TaskViapoint, which implements a Task in which the first N columns in cost_vars should represent a N-D trajectory. This convention is respected by TaskSolverDmp, which is able to generate such trajectories.
-
-
-EigenToString.tpp from EigenBoostSerialization? Or move into EigenFileIO
