+++ demo_robot/step1B_trainDmpFromTrajectoryFilePlot.py
@@ -29,7 +29,7 @@ import matplotlib.pyplot as plt
 from mpl_toolkits.mplot3d import Axes3D
 import os, sys
 
-lib_path = os.path.abspath('../../python')
+lib_path = os.path.abspath('../python')
 sys.path.append(lib_path)
 from functionapproximators.functionapproximators_plotting import * 
 from dmp.dmp_plotting import * 
