+++ demo_robot/step4C_plotOptimization.py
@@ -1,6 +1,6 @@
 # This file is part of DmpBbo, a set of libraries and programs for the 
 # black-box optimization of dynamical movement primitives.
-# Copyright (C) 2018 Freek Stulp
+# Copyright (C) 2014 Freek Stulp, ENSTA-ParisTech
 # 
 # DmpBbo is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Lesser General Public License as published by
@@ -18,42 +18,32 @@
 
 import os
 import sys
-import glob
 import pickle
 import numpy as np
 import matplotlib.pyplot as plt
 
-lib_path = os.path.abspath('../../python/')
+lib_path = os.path.abspath('../python')
 sys.path.append(lib_path)
 
+from dmp_bbo.Task import Task
 from dmp_bbo.dmp_bbo_plotting import *
-from dmp.dmp_plotting import *
-
 
 if __name__=="__main__":
 
-    input_cost_vars_file = None
-    input_task_file = None
-    
-    if len(sys.argv)<3:
-        print('Usage: '+sys.argv[0]+' <cost vars file> <task pickle file>')
-        print('Example: python3 '+sys.argv[0]+' results/cost_vars_demonstrated.txt results/task.p')
+    # See if input directory was passed
+    if (len(sys.argv)>=2):
+        directory = str(sys.argv[1])
+    else:
+        print('\nUsage: '+sys.argv[0]+' <directory>\n')
         sys.exit()
-        
-    if (len(sys.argv)>1):
-        input_cost_vars_file = sys.argv[1]
-    if (len(sys.argv)>2):
-        input_task_file = sys.argv[2]
-        
-    task = None
-    if input_task_file:
-        task = pickle.load(open(input_task_file, "rb" ))
-    
-    fig = plt.figure(1)
-    n_subplots = 1
-    axs = [ fig.add_subplot(1,n_subplots,ii+1) for ii in range(n_subplots) ]
-        
-    cost_vars = np.loadtxt(input_cost_vars_file)
-    task.plotRollout(cost_vars,axs[0])
 
+    # Load the task
+    task = pickle.load(open(directory+'/task.p', "rb" ))
+    
+    # Plot the optimization results (from the files saved to disk)
+    golden_ratio = 1.618 # graphs have nice proportions with this ratio
+    n_subplots = 4
+    fig = plt.figure(3,figsize=(golden_ratio*n_subplots*3,3))
+    plotOptimizationRolloutsTask(directory,fig,task)
+    plt.tight_layout()
     plt.show()
