+++ demo_robot/step4A_oneOptimizationUpdate.py
@@ -22,7 +22,7 @@ import pickle
 import numpy as np
 import matplotlib.pyplot as plt
 
-lib_path = os.path.abspath('../../python')
+lib_path = os.path.abspath('../python')
 sys.path.append(lib_path)
 
 from bbo.updaters import *
