+++ demo_robot/TaskThrowBall.py
@@ -3,7 +3,7 @@ import sys
 import numpy as np
 import matplotlib.pyplot as plt
 
-lib_path = os.path.abspath('../../python')
+lib_path = os.path.abspath('../python')
 sys.path.append(lib_path)
 
 from dmp_bbo.Task import Task
