### 2019-04-22 1749958
Added namespaces std:: and Eigen:: to cpp files in functions
This does not change the functionality of the code, but enables Doxygen to associate functin prototypes in cpp and hpp with eachother.
