### 2019-04-22 2b09b0b
Doxygen documentation added for several functions
