+++ src/dmp/DmpWithGainSchedules.hpp
@@ -71,28 +71,53 @@ class Trajectory;
 /**
 Implementation of DMPs which contain extra dimensions to represent variable gain schedules, as described in \cite buchli11learning.
 
-These dimensions can also used to represent force profiles \cite kalakrishnan11learning, or any other variables relevant to the Dmp, but which are not part of the dynamical system.
+These dimensions can also used to represent force profiles \cite kalakrishnan11learning, or any other variables relevant to the DMP, but which are not part of the dynamical system.
  */
 class DmpWithGainSchedules : public Dmp
 {
 public:
   
- DmpWithGainSchedules(
-   Dmp* dmp,
-   std::vector<FunctionApproximator*> function_approximators_gain_schedules
-   );
+  /** Constructor.
+   * \param[in] dmp The Dmp part of the DmpWithGainSchedules
+   * \param[in] function_approximators_gain_schedules Function approximators that will represent the gain schedules.
+   */
+  DmpWithGainSchedules(
+    Dmp* dmp,
+    std::vector<FunctionApproximator*> function_approximators_gain_schedules
+  );
    
   
   /** Destructor. */
   ~DmpWithGainSchedules(void);
   
   /** Return a deep copy of this object 
-   * \return A deep copy of this object
+   * \return A deep copy of this object    
    */
   DmpWithGainSchedules* clone(void) const;
   
+  /** Start integrating the system
+   *
+   * \param[out] x     - The first vector of state variables
+   * \param[out] xd    - The first vector of rates of change of the state variables
+   * \param[out] gains - The gains of the gain schedules
+   *
+   * \remarks x, xd, and gains should be of size dim() X 1. This forces you to pre-allocate 
+   * memory, which speeds things up (and also makes Eigen's Ref functionality easier to deal with).
+   */
   void integrateStart(Eigen::Ref<Eigen::VectorXd> x, Eigen::Ref<Eigen::VectorXd> xd, Eigen::Ref<Eigen::VectorXd> gains) const;
 
+  /**
+   * Integrate the system one time step.
+   *
+   * \param[in]  dt         Duration of the time step
+   * \param[in]  x          Current state
+   * \param[out] x_updated  Updated state, dt time later.
+   * \param[out] xd_updated Updated rates of change of state, dt time later.
+   * \param[out] gains      The gains of the gain schedules
+   *
+   * \remarks x should be of size dim() X 1. This forces you to pre-allocate memory, which
+   * speeds things up (and also makes Eigen's Ref functionality easier to deal with).
+   */
   void integrateStep(double dt, const Eigen::Ref<const Eigen::VectorXd> x, Eigen::Ref<Eigen::VectorXd> x_updated, Eigen::Ref<Eigen::VectorXd> xd_updated, Eigen::Ref<Eigen::VectorXd> gains) const;
   
   /**
@@ -134,6 +159,10 @@ public:
    */
   void train(const Trajectory& trajectory, std::string save_directory, bool overwrite=false);
   
+  /** 
+   * Return the dimensionality of the vector with gains.
+   * \return The dimensionality of the vector with gains
+   */
   int dim_gains(void) const
   {
     return function_approximators_gains_.size();
