### 2019-04-22 bb947ee
Better explanation of Python/C++ use cases.
