### 2019-04-22 56a618e
Updated doxygen file with "doxygen -u"
Necessary because some flags are now deprecated, which lead to warnings.
