### 2019-04-22 51b3b98
Removed LaTeX version of tutorial. Available in tutorials/*.md now.
