+++ docs/tutorial/tutorial.tex
@@ -1,6 +1,0 @@
-\documentclass{article}
-\usepackage{hyperref}
-
-\begin{document}
-This tutorial has moved to \url{http://github.com/stulp/dmpbbo/tutorial}
-\end{document}
