### 2019-04-22 e18796b
Renamed demos/ to demos_cpp/ for consistency.
