+++ CMakeLists.txt
@@ -69,7 +69,7 @@ include_directories ( ${Boost_INCLUDE_DIRS} )
 ###############################################################################
 # Things to compile are in the src/ directory
 add_subdirectory(src)
-add_subdirectory(demos)
+add_subdirectory(demos_cpp)
 add_subdirectory(demo_robot)
 
 ###############################################################################
