### 2019-04-22 50358e8
Fixed links to tutorials
