+++ src/functionapproximators/FunctionApproximator.hpp
@@ -359,7 +359,7 @@ namespace DmpBbo {
 
 /** \page page_func_approx Function Approximation
 
-This page provides and overview of the implementation of function approximators in the \c dynamicalsystems/ module.
+This page provides an  overview of the implementation of function approximators in the \c dynamicalsystems/ module.
 
 It is assumed you have read about the theory behind function approximators in the tutorial <a href="https://github.com/stulp/dmpbbo/tutorial/functionapproximators.md">tutorial/functionapproximators.md</a>.
 
