### 2019-04-22 d6ec4b3
Fixed links to tutorials in tutorial/
