+++ src/dynamicalsystems/DynamicalSystem.hpp
@@ -413,7 +413,7 @@ namespace DmpBbo {
 
 This page provides an  overview of the implementation of dynamical systems in the \c dynamicalsystems/ module.
 
-It is assumed you have read about the theory behind dynamical systems in the tutorial <a href="https://github.com/stulp/dmpbbo/tutorial/dynamicalsystems.md">tutorial/dynamicalsystems.md</a>.
+It is assumed you have read about the theory behind dynamical systems in the tutorial <a href="https://github.com/stulp/dmpbbo/blob/master/tutorial/dynamicalsystems.md">tutorial/dynamicalsystems.md</a>.
 
 \section sec_dyn_sys_analytical_solution Analytical solution of a dynamical system
 
