### 2019-04-22 afb4ead
Fixed problem with Eigen assigment.
This compiled on my Ubuntu 16.04 installation, but not on travis.
