+++ demo_robot/README.md
@@ -1,3 +1,1 @@
-# Demos and tips for running dmpbbo on a real robot
-
-Please see the tutorial on this: http://github.com/stulp/dmpbbo/tutorial/dmp_bbo_robot.md
+../tutorial/dmp_bbo_robot.md
