### 2019-04-22 bc6e3e7
Moved md file to be inside the demo_robot directory (closer to where the illustrated programs are)
