### 2019-04-22 #1225 @stulp
@danielskatz - I've just merged the branches in which I was addressing the remaining issues.

My main change has been to completely rewrite the code in demos/dmp_bbo_robot (and renaming it to demo_robot/). The directory https://github.com/stulp/dmpbbo/blob/master/demo_robot has been the result of that, and the programs/scripts/README in that directory have not been reviewed yet. I think this should fix the remaining issue #44 raised by @studywolf

Along the way, I made several other fixes:
* Fixing boost:serialization issues.
* Fixed all doxygen warnings (by documenting all functions and adding namespaces)
* Nicer plotting here and there (to make nice plots for demo_robot/)
These fixes/tweaks, I would say, do not require a review.

I will make a new release with a new version number and an archive tomorrow.
