### 2019-04-22 #1225 @danielskatz
👋 @studywolf - I know you thought you were done, but I would greatly appreciate your confirmation that the submission, with these changes, is still acceptable (and ideally has improved).
