### 2019-04-22 #1225 @stulp
The new release is here: https://github.com/stulp/dmpbbo/releases/tag/v1.0.0

Archives of the repository are downloadable at the end of that release.
