### 2019-05-03 #1225 @studywolf
Hello! Yes you are correct, sorry about that. I've run through the demo_robot files and just made a typo correction issue in the README. This all LGTM :+1:
