### 2019-05-06 #1225 @danielskatz
Ok. I jumped the gun.

@stulp - please deposit the current version in an archival repository, eg zenodo, figshare, etc and let me know the DOI. And let me know the version number that goes with the archived software.
