### 2019-05-06 456b2aa
Typo in the README.
Fixes #47
