### 2019-05-06 #1225 @graiola
@danielskatz this is the DOI 10.5281/zenodo.2669652
The version is v1.0.0

Thanks!
