### 2019-05-06 #1225 @danielskatz
whedon set 1.0.0 as version
