### 2019-05-06 #1225 @danielskatz
👋 @graiola - I no longer see the paper.md and paper.bib files in the repo - what am I missing?
