### 2019-05-07 #1225 @arfon
@danielskatz - they're still there (e.g. `https://github.com/stulp/dmpbbo/blob/master/docs/paper/paper.md`). Whedon just sometimes fails to clone the repositories properly (a long-standing bug).
