### 2019-05-07 7c15c71
fixing references for JOSS paper
