+++ docs/dmp.bib
@@ -9,7 +9,7 @@
   number                   = {7},
   pages                    = {820-833},
   volume                   = {30},
-  doi                      = {https://doi.org/10.1177/0278364911402527},
+  doi                      = {10.1177/0278364911402527},
 
   abstract                 = {One of the hallmarks of the performance, versatility, and robustness of biological motor control is the ability to adapt the impedance of the overall biomechanical system to different task requirements and stochastic disturbances. A transfer of this principle to robotics is desirable, for instance to enable robots to work robustly and safely in everyday human environments. It is, however, not trivial to derive variable impedance controllers for practical high degree-of-freedom (DOF) robotic tasks. In this contribution, we accomplish such variable impedance control with the reinforcement learning (RL) algorithm PI2 ({\bf P}olicy {\bf I}mprovement with {\bf P}ath {\bf I}ntegrals). PI2 is a model-free, sampling based learning method derived from first principles of stochastic optimal control. The PI2 algorithm requires no tuning of algorithmic parameters besides the exploration noise. The designer can thus fully focus on cost function design to specify the task. From the viewpoint of robotics, a particular useful property of PI2 is that it can scale to problems of many DOFs, so that reinforcement learning on real robotic systems becomes feasible. We sketch the PI2 algorithm and its theoretical properties, and how it is applied to gain scheduling for variable impedance control. We evaluate our approach by presenting results on several simulated and real robots. We consider tasks involving accurate tracking through via-points, and manipulation tasks requiring physical contact with the environment. In these tasks, the optimal strategy requires both tuning of a reference trajectory \emph{and} the impedance of the end-effector. The results show that we can use path integral based reinforcement learning not only for planning but also to derive variable gain feedback controllers in realistic scenarios. Thus, the power of variable impedance control is made available to a wide variety of robotic systems and practical applications.},
   url                      = {http://ijr.sagepub.com/content/early/2011/03/31/0278364911402527}
@@ -23,21 +23,21 @@
   number                   = {2},
   pages                    = {328--373},
   volume                   = {25},
-  doi                      = {https://doi.org/10.1162/neco_a_00393},
+  doi                      = {10.1162/neco_a_00393},
 }
 
 @InProceedings{ijspeert02movement,
   title                    = {Movement imitation with nonlinear dynamical systems in humanoid robots},
   author                   = {A. J. Ijspeert and J. Nakanishi and S. Schaal},
-  booktitle                = {Proceedings of the IEEE International Conference on Robotics and Automation (ICRA)},
-  doi                      = {https://doi.org/10.1109/ROBOT.2002.1014739},
+  booktitle                = {Proceedings of the {IEEE} International Conference on Robotics and Automation ({ICRA})},
+  doi                      = {10.1109/ROBOT.2002.1014739},
   year                     = {2002},
 }
 
 @InProceedings{kalakrishnan11learning,
   title                    = {Learning force control policies for compliant manipulation},
   author                   = {Kalakrishnan, M. and Righetti, L. and Pastor, P. and Schaal, S.},
-  booktitle                = {IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS 2011)},
+  booktitle                = {{IEEE/RSJ} International Conference on Intelligent Robots and Systems ({IROS} 2011)},
   year                     = {2011},
 
   url                      = {http://www-clmc.usc.edu/publications/K/kalakrishnan-IROS2011}
@@ -51,7 +51,7 @@
   number                   = {1},
   pages                    = {145-157},
   volume                   = {28},
-  doi                      = {https://doi.org/10.1109/tro.2011.2163863},
+  doi                      = {10.1109/tro.2011.2163863},
 
[…]
