### 2019-05-07 #1225 @danielskatz
@graiola - Some of the references have some issues - please see https://github.com/stulp/dmpbbo/pull/49
