### 2019-05-07 #1225 @danielskatz
now we wait for the DOI to work, then I'll complete the process
