### 2019-05-07 #1225 @arfon
> now we wait for the DOI to work, then I'll complete the process

@danielskatz - the DOI is now resolving.
