### 2019-05-08 #1225 @stulp
Thank you @studywolf and @jenskober for your thorough reviews, and thank you @arfon and @danielskatz for supervising the review process. I think the reviewing process has certainly contributed to improving the code, especially in terms of documentation, and making the code easier to navigate for newcomers. So thanks again for the time and effort you have all invested.
