### 2019-05-08 #1225 @jenskober
Yes, it felt more like helping to improve the code rather than typical academic reviews. I really enjoyed the process, especially as @graiola and @stulp were extremely quick to react and went beyond what we were suggesting.
